+++ package.json
@@ -20,12 +20,8 @@
 		"@sentry/react": "^9.1.0",
 		"@sentry/vite-plugin": "^3.2.0",
 		"html2pdf.js": "^0.10.2",
-		"i18next": "^23.16.8",
-		"i18next-browser-languagedetector": "^8.0.0",
-		"i18next-http-backend": "^3.0.1",
 		"lodash": "^4.17.21",
 		"preact": "^10.22.1",
-		"preact-i18next": "^11.8.5",
 		"preact-iso": "^2.6.3",
 		"react": "npm:@preact/compat",
 		"react-calendar": "^5.1.0",
