+++ src/components/Header/NavigationMenu.tsx
@@ -2,53 +2,10 @@ import { useLocation } from 'preact-iso';
 import { useState } from 'preact/hooks';
 import { MenuList, MenuListItem, StyledButton } from "react95";
 import styled from 'styled-components';
-import { Trans } from '@lingui/react/macro';
-import { msg, t } from '@lingui/core/macro';
-import { MessageDescriptor, i18n } from '@lingui/core';
+import { Trans, useLingui } from '@lingui/react/macro';
 
 import DownGreenTri from '../../assets/down-green-tri.png';
-
-export interface MenuConfigItem {
-    name: string;
-    labelMsg?: MessageDescriptor;
-    url: string;
-    img: string;
-}
-
-export const MenuConfig: MenuConfigItem[] = [
-    {
-        name: t`Home`,
-        labelMsg: msg({
-            message: 'Home'
-        }),
-        url: '/',
-        img: '🏠',
-    },
-    {
-        name: t`About`,
-        labelMsg: msg({
-            message: 'About'
-        }),
-        url: '/about',
-        img: '📖',
-    },
-    {
-        name: t`The Crucible`,
-        labelMsg: msg({
-            message: 'The Crucible'
-        }),
-        url: '/crucible',
-        img: '🫠',
-    },
-    {
-        name: t`Summary`,
-        labelMsg: msg({
-            message: 'Summary'
-        }),
-        url: '/summary',
-        img: '📊',
-    },
-];
+import { MenuConfig } from './MenuConfig';
 
 const StyledMenuList = styled(MenuList)`
     list-style: none;
@@ -102,6 +59,12 @@ const IconWrapper = styled.span.attrs<any>({
 
 function NavigationMenu(){
     const [open, setOpen] = useState(false);
+    const { i18n } = useLingui();
+
+    if (!i18n || !i18n.locale) {
+        return null;
+    }
+
     const { url } = useLocation();
 
     return (
