--- src/i18n/i18n.ts
@@ -1,28 +1,0 @@
-import i18n from 'i18next';
-import { initReactI18next } from 'preact-i18next';
-import LanguageDetector from 'i18next-browser-languagedetector';
-import Backend from 'i18next-http-backend';
-
-i18n
-    .use(Backend)
-    .use(LanguageDetector)
-    .use(initReactI18next)
-    .init({
-    fallbackLng: 'en',
-    debug: true,
-    interpolation: {
-        escapeValue: false,
-    },
-        resources: {
-            en: {
-                translation: {
-                    specialty: {
-                        one: 'specialty',
-                        other: 'specialties'
-                    }
-                }
-            }
-        }
-    });
-
-export default i18n;
