+++ src/locales/de/messages.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"stat.reminder\":[[\"reminder\"]],\"stat.label\":[[\"stat\"]],\"stat.intelligence.reminder\":[\"Cunning, logic, and intuition\"],\"stat.constitution.reminder\":[\"Health and hardiness\"],\"stat.charisma.reminder\":[\"Personal and physical appeal\"],\"stat.strength.reminder\":[\"Raw physical power\"],\"stat.dexterity.reminder\":[\"Speed and physical agility\"],\"stat.power.reminder\":[\"Will, spirit, and mental stability\"],\"8VjWl0\":[\"--- Additional Professions ---\"],\"KE5vnC\":[\"--- Standard Professions ---\"],\"7o9kk2\":[\"(Choose type below)\"],\"J/hVSQ\":[[\"0\"]],\"X3V8cR\":[[\"0\"]],\"xYxQCZ\":[[\"0\"],\" \",[\"1\"]],\"Zz4Qvw\":[[\"capitalizedStat\"]],\"w9dLMW\":[[\"label\"]],\"JpzPnP\":[[\"labelText\"]],\"choose-skills.no-choices\":[[\"professionName\"],\" has no flexible skills to choose from.\"],\"leNkYU\":[[\"reminderText\"]],\"5ft2yB\":[[\"skillCapText\"]],\"kDmBhk\":[[\"statKey\"]],\"urTZDx\":[[\"statKey\"]],\"LZJAen\":[\"`$\",[\"0\"],\" has no flexible skills to choose from.`\"],\"9eGA8Z\":[\"010111100001\"],\"Hzaa8H\":[\"1. Last Name, First Name, Middle Initial\"],\"yQJwH3\":[\"10. Physical Description\"],\"kmhPgQ\":[\"11. Bonds\"],\"yGWS1j\":[\"12. Motivations and Mental Disorders\"],\"w/Xy4v\":[\"13. Incidents of San Loss Without Going Insane\"],\"VnPSmP\":[\"2. Profession (Rank if Applicable)\"],\"qVQ1QT\":[\"3. Employer\"],\"nVobO+\":[\"4. Nationality\"],\"icgCRe\":[\"5. Sex\"],\"+18IK1\":[\"6. Age and D.O.B.\"],\"AbZcNQ\":[\"7. Education and Occupational History\"],\"bzqTDJ\":[\"8. Statistics\"],\"W5FXNn\":[\"9. Derived Attributes\"],\"5XYRnE\":[\"A dedicated hobbyist; with a foreign language you can have rudimentary conversations.\"],\"igi2jT\":[\"A lifetime’s pursuit or multiple doctorates.\"],\"uyJsf6\":[\"About\"],\"2zPUa5\":[\"About the Author\"],\"UNDRet\":[\"About the Game\"],\"WcvywZ\":[\"Accounting\"],\"F6pfE9\":[\"Active\"],\"zq0AxB\":[\"Active Couch Potato\"],\"yGwMqg\":[\"adapted\"],\"0kRNUo\":[\"Add +10% to your Agent's Occult and +10% to any five skills other than Unnatural. This can bring no skill higher than 90%. Reduce your Agent's SAN by 5. Remove one Bond.\"],\"Xi6Is8\":[\"Add +10% to your Agent's Occult skill. Reduce SAN by 5. Subtract 3 from your Agent's CHA and each Bond. Your Agent is adapted to violence (see page 73).\"],\"EHRBpm\":[\"Add +10% to your Agent's Occult skill. Reduce SAN by 5. Subtract 3 from your Agent's POW. Your Agent is adapted to helplessness (see page 73).\"],\"custom-skill-form.bonus-points-reminder\":[\"Adds \",[\"DEFAULT_BONUS_VALUE\"],\" to the skill total for each bonus point allocated (capped at \",[\"DEFAULT_TOTAL_CAP\"],\")\"],\"7TlSPN\":[\"Adroit\"],\"zVgowR\":[\"Affable\"],\"CmUtio\":[\"Agile\"],\"gXy38H\":[\"Air or sea, commercial or military, your duty is to keep your passengers alive and craft intact. This can lead to hard choices when your passengers put the vehicle in danger. Or are you a drone operator, flying a Predator from a thousand miles away? Either way, what op brought you to the attention of Delta Green?\"],\"Z+c4we\":[\"Aircraft\"],\"rPKlAj\":[\"Alertness\"],\"EbuTi3\":[\"Alias\"],\"JIeg5e\":[\"All Thumbs\"],\"6zd6CL\":[\"Alluring\"],\"UCIiuP\":[\"Analysing thought\"],\"WbrlFU\":[\"Andrew Sobiesiak is a Software Engineer that has played TTRPGs online weekly since 2017.<0/> He would make a terrible field agent but is prepared to fight the sisyphean cosmic war. He Lives in <1>Seattle, WA</1> with his <2>wife, 5 cats, and a dog</2>\"],\"2qgwCU\":[\"Anthropologist\"],\"juHsSO\":[\"Anthropology\"],\"aAIQg2\":[\"Appearance\"],\"f+WR1d\":[\"Applicable Skill Sets\"],\"FGdRNI\":[\"Apply preset skills and choose additional skills for your character.\"],\"xLXsjU\":[\"Arc Dream Publishing is a small-press publisher that produces high-quality games, including roleplaying games, a horror gaming magazine, and card games. Some of their popular works include Delta Green, Swords & Sorceries, and The Black Company RPG.\"],\"vJo5g6\":[\"Archeology\"],\"+QARA4\":[\"Art\"],\"Zvntde\":[\"Art (Painting)\"],\"ziY9j1\":[\"Artillery\"],\"3Vj2Gv\":[\"As a member of the United States Marine Corps, you are part of America's elite expeditionary force in readiness. First to fight, you've been trained to operate in any environment, from urban warfare to amphibious operations. Your intense training and esprit de corps set you apart. Whatever caught Delta Green's attention, it wasn't your first brush with hell.\"],\"EOS+bp\":[\"As part of a force like the U.S. Army Rangers, you volunteered for a more difficult path than other soldiers. You've spent years in the most grueling training on the planet, and now serve on the most dangerous missions around.\"],\"vS6YOc\":[\"Athletic\"],\"JGpQgX\":[\"Athletics\"],\"BO9EJL\":[\"Avoiding getting hit\"],\"GDkUlj\":[\"Awkward\"],\"ZuyHGl\":[\"Base Statistics\"],\"o5DA10\":[\"Beliefs\"],\"Bj0YCx\":[\"Bland\"],\"pFWzUl\":[\"Bond description\"],\"joDnA9\":[\"Bond Detail\"],\"Qp5ZKm\":[\"Bond Name\"],\"004WwA\":[\"Bond Type\"],\"MTMViL\":[\"Bonds\"],\"ooT9hB\":[\"Bonds Details\"],\"YfLs0q\":[\"Bonds represent meaningful relationships your agent has with non-player characters.\"],\"7zP6Gy\":[\"Bonds represent your Agent's relationships with other people.\\n                    Each bond starts with a score equal to your Agent's Charisma (\",[\"0\"],\").\\n                    The score of a bond can improve if it is nurtured between operations. A bond can never have a score higher than your Agent's Charisma.\\n                    Any time your Agent's Charisma drops, each bond score drops by the same amount.\\n                    Bonds can be defined as a single person or a group of people. For example, an agent might have bonds with their spouse, their business partner, and their bowling league.\\n                    A Bond must be a real person or small group of people who are alive and can be interacted with.\"],\"PqoQ2W\":[\"Bonus\"],\"Mxgo1h\":[\"Bonus Points\"],\"jw0SVP\":[\"Bonus Points Remaining\"],\"MIHH6C\":[\"Breaking Point\"],\"jEvJsi\":[\"Bright\"],\"GtWnxo\":[\"Brilliant\"],\"Nr0SIL\":[\"Bureaucracy\"],\"CBGYlg\":[\"Business Math\"],\"6rScUj\":[\"Camping, tracking, improvising in nature\"],\"a46V7b\":[\"Captivating\"],\"aw5Ht8\":[\"Captivity or Imprisonment\"],\"5NsOuM\":[\"Cat-like\"],\"hVHY2j\":[\"Charisma\"],\"gLHd7f\":[\"Check a Bond’s box when projecting sanity damage\"],\"DK/qpg\":[\"Check a skill when you fail a skill check. After a session, add 1d4 to each checked skill and erase the check.\"],\"nUIcQr\":[\"Choose \",[\"0\"],\" Additional Skills\"],\"zikS7b\":[\"Choose a bonus skill package\"],\"k/xMy9\":[\"Choose a package to automatically apply bonus points, or select 'No Package' to manually assign your bonus points\"],\"CGI4Ui\":[\"Church or support group\"],\"xCJdfg\":[\"Clear\"],\"Jz0/oR\":[\"Clear Descriptors\"],\"LndrJY\":[\"Clear Selected Skills\"],\"Ql//nl\":[\"Clear Templates\"],\"OxOc2C\":[\"Clumsy\"],\"RLWPXk\":[\"Colleagues in an intense job\"],\"KLEQcs\":[\"College major or years of experience; with 50% in a foreign language you have native fluency.\"],\"rIdFzg\":[\"College minor or basic training.\"],\"yKoSLY\":[\"Compelling\"],\"kHbyNa\":[\"Computer Science\"],\"AU0Kdy\":[\"Computers and machinery are the backbone of modern industry. You are a craftsman with data or machinery, possibly for the government and most definitely for profit. However you use your skills, the overlap between information technology and awareness of the unnatural could make this the most dangerous job on the planet.\"],\"fOaHdd\":[\"Constitution\"],\"uhhJuh\":[\"Controlled explosions\"],\"IbtVsv\":[\"Convincing people\"],\"3zF2gs\":[\"Couch Potato\"],\"gtWP1l\":[\"Could take the stairs\"],\"A8wadO\":[\"Coworker or partner\"],\"UTfQwZ\":[\"Craft\"],\"470Fk4\":[\"Craft (Macrame)\"],\"XjzLm3\":[\"Creative pursuits\"],\"PMunmF\":[\"Criminal\"],\"oWJvZ7\":[\"Criminology\"],\"VoIJ7Q\":[\"CSI\"],\"5ALalD\":[\"Cult shit\"],\"d/rSvQ\":[\"Cunning, logic, and intuition\"],\"Hp1l6f\":[\"Current\"],\"zIo0DE\":[\"Custom Professions\"],\"7/D6Vk\":[\"Dabbler.\"],\"iX6K3g\":[\"Damaged Veteran Templates\"],\"7uHWR+\":[\"Date of Birth\"],\"onAw8T\":[\"Decades of experience, or a graduate or doctoral degree.\"],\"tffdBt\":[\"Declassified Record\"],\"L3rQ+j\":[\"Deft\"],\"VVywaC\":[\"Delicate\"],\"PgSdTp\":[\"Demolitions\"],\"ezghSE\":[\"Derived Attributes\"],\"Nu4oKW\":[\"Description\"],\"s3Uzfx\":[\"Determined\"],\"08z1KE\":[\"Dexterity\"],\"u9l1aF\":[\"Dice\"],\"DwvMe2\":[\"Dim\"],\"BJ6YQv\":[\"Disguise\"],\"uPWSEh\":[\"Distinguishing Features\"],\"9/Nzrk\":[\"Do you close your eyes to what you’ve seen and go back to sleep? Or do you come with this psycho burnout and do the impossible against the unbelievable and keep the future at bay for another day?\"],\"difbU5\":[\"Do you like drugs?\"],\"DM17Gg\":[\"Doctors are often the first to uncover signs of an unnatural incursion, and the most valuable investigators of its disastrous effects on humanity.\"],\"p+tvhp\":[\"Dodge\"],\"OUmX26\":[\"Dossier\"],\"Q2ObDC\":[\"Drive\"],\"1afgyS\":[\"Dull\"],\"aqxYLv\":[\"Education\"],\"n9u/m0\":[\"Employer\"],\"pbUVnA\":[\"Engaging\"],\"iy5fSS\":[\"Engineer\"],\"RI/fzG\":[\"Enter a Subtype\"],\"HmI5oK\":[\"Example\"],\"O4wxl8\":[\"Export as PDF\"],\"S8GwkS\":[\"Extreme Violence\"],\"F7dRcr\":[\"Fair\"],\"+UN+Ez\":[\"Fatigued\"],\"y+uhgI\":[\"Federal Agent\"],\"hrDJ2/\":[\"Feeble\"],\"oCrptp\":[\"Finding the path\"],\"qUl/bH\":[\"Finding things\"],\"GnG+vN\":[\"Firearms\"],\"/OsYEc\":[\"Firefighter\"],\"RtEHo7\":[\"Firm\"],\"11+tdk\":[\"First Aid\"],\"kODvZJ\":[\"First Name\"],\"ZX5IeB\":[\"First Night at the Opera?\"],\"u4iXQi\":[\"Fit\"],\"wMLcWn\":[\"Flexible in principle\"],\"yaw+7x\":[\"Font Settings\"],\"kXTwNy\":[\"Foreign Languages\"],\"aFFQEb\":[\"Foreign Languages (French)\"],\"a9k1oC\":[\"Foreign Service Officer\"],\"z7q/DL\":[\"Forensics\"],\"K9cGgQ\":[\"Forgettable\"],\"0BlDs+\":[\"Forklifts, cranes, excavators, etc\"],\"bdeAb2\":[\"Formidable\"],\"noN438\":[\"Fortified\"],\"1Yrw83\":[\"Found on page \",[\"pageNumber\"]],\"R2sFu3\":[\"Fragile\"],\"ZQfuYY\":[\"Frail\"],\"nLC6tu\":[\"French\"],\"yTcD6q\":[\"Fumbling\"],\"ZJ1pST\":[\"Gene Parmesan, he's the best\"],\"7o4S1E\":[\"Genius\"],\"+/PqzT\":[\"Graceless\"],\"47ofD/\":[\"Greased Lightning\"],\"fcFvf1\":[\"Greasing the wheels of government\"],\"L8fEEm\":[\"Group\"],\"J36o9O\":[\"Handy\"],\"fjd2D3\":[\"Hard Experience\"],\"bzM4ak\":[\"Health and hardiness\"],\"yGS9cI\":[\"Healthy\"],\"UjSOv0\":[\"Hearty\"],\"NxZ7K+\":[\"Heavy Machinery\"],\"Nw0C1v\":[\"Heavy Weapons\"],\"0ZBnED\":[\"Helplessness\"],\"ORe4Lw\":[\"Historian\"],\"0caMy7\":[\"History\"],\"Bwb/M2\":[\"Hit Points\"],\"1+WSY1\":[\"Hobbies\"],\"i0qMbr\":[\"Home\"],\"UU7+rG\":[\"Horses and such\"],\"21qVcE\":[\"Huge\"],\"Po1nwb\":[\"HUMINT\"],\"4bNpE0\":[\"If the intelligence community was a family, Delta Green would be the uncle nobody talks about.<0/>Delta Green is a Table Top Role Playing Game that uses a D100 system. Handlers run operations for their agents, other players, where a grey government operation works to protect the world from unkown cosmic horrors.\"],\"dbLUe7\":[\"If you haven't studied it you're doomed to repeat it\"],\"p64A7d\":[\"If you want to support this work, you can buy some Delta Green content through affiliate links.\"],\"NNy6W8\":[\"Impulsive\"],\"b23dA7\":[\"In the FBI, NSA and CIA, there are those who gather information and those who decide what it means. You take information from disparate sources—newspapers, websites, informants, ELINT, and the assets developed by Case Officers—and figure out what it means. In short, your job is the piecing together of unrelated knowledge, a dangerous endeavor in the world of Delta Green.\"],\"UFi3x2\":[\"Incurious\"],\"fByw/g\":[\"Individual\"],\"+RRbXc\":[\"Indomitable\"],\"aemL3D\":[\"Infirmed\"],\"C5kxqk\":[\"Inner Adam Savage\"],\"s2/OPX\":[\"Insipid\"],\"Gzw2pq\":[\"Intelligence\"],\"Mf5ErF\":[\"Intelligence Analyst\"],\"3FDXvB\":[\"Intelligence Case Officer\"],\"VGX26E\":[\"Irrational\"],\"Jp7JUr\":[\"It's a jeep thing, you wouldn't understand\"],\"mEzWf0\":[\"Jacked\"],\"tcBJo7\":[\"Jumping, ducking, running, climbing, etc\"],\"yoMIHR\":[\"Jurassic Park\"],\"sNoMVY\":[\"Key Stats\"],\"zC0r/K\":[\"Knives, hatchets, swords, etc\"],\"87a/t/\":[\"Label\"],\"ok/7qO\":[\"Land\"],\"UXBCwc\":[\"Last Name\"],\"iqhOTb\":[\"Law\"],\"whBNst\":[\"Lawyer\"],\"XHdFzA\":[\"List character motivations and any mental disorders...\"],\"AqDRRD\":[\"LMGs I think\"],\"1z0Tzf\":[\"Macrame\"],\"kT7U6s\":[\"Magnetic\"],\"oLEThm\":[\"Making things explode from far away\"],\"PONy2U\":[\"Manual Input\"],\"LlSUa5\":[\"Many Delta Green Agents are federal law enforcement officers, mostly from the FBI. Delta Green decided long ago that federal agents have the optimum balance of skills and mental stability needed to confront the unnatural.\"],\"PAXVdT\":[\"Marine\"],\"CK1KXz\":[\"Max\"],\"spK3cv\":[\"Media Specialist\"],\"zeVoXl\":[\"Medical practice\"],\"hS5cj7\":[\"Medical professionals are on the front line when awful things happen. Is that what brought you to the group's attention?\"],\"6Kw9Gv\":[\"Medicine\"],\"F5rArt\":[\"Mediocre\"],\"syonIH\":[\"Meet me at <0>The Crucible</0> and tell me a little bit about yourself, agent. What's your deal?\"],\"qx+7SQ\":[\"Melee Weapons\"],\"e3Oj4X\":[\"Mesmerizing\"],\"2NOxz2\":[\"Middle Initial\"],\"dImRx9\":[\"Military Science\"],\"YOtRZ0\":[\"Military Science (Land)\"],\"hNi0vh\":[\"Military tactics\"],\"lJ71dZ\":[\"Minor urgent medical care\"],\"dOLpsA\":[\"Motivations\"],\"aijpeY\":[\"Moving oneself through water\"],\"e9CT52\":[\"Muddying the waters of a crime-scene\"],\"av28MK\":[\"Muscular\"],\"2YKBHm\":[\"Nationality\"],\"dbDjdV\":[\"Navigate\"],\"mqUHTq\":[\"Nigh Immortal\"],\"Ysyt+a\":[\"Nimble\"],\"kSzR8n\":[\"No Package (Manual Bonus Points)\"],\"u3hw5g\":[\"Note additional details about the bond\"],\"XOGTWH\":[\"Noticing things\"],\"9OGltc\":[\"Number of social connections available to your character.\"],\"tEPJOu\":[\"Nurse\"],\"xhyJlX\":[\"Obsessions\"],\"7JOZLS\":[\"Occult\"],\"i0qI4Z\":[\"Opearting a motor-vehicle\"],\"ZI3xDG\":[\"Operating flying vehicles\"],\"0zpgxV\":[\"Options\"],\"DW8nOh\":[\"Oracular\"],\"08AP9H\":[\"Painting\"],\"pYwzJ9\":[\"Parles-vous Français?\"],\"G/Cl3b\":[\"Pays for Mensa\"],\"Xvtlo9\":[\"Perserverant\"],\"iLp4WI\":[\"Personal and physical appeal\"],\"DhAC6+\":[\"Personal Data\"],\"sJYPeP\":[\"Personal Details\"],\"8peLId\":[\"Personal Motivation \",[\"0\"]],\"VpAfxU\":[\"Personal Motivations\"],\"39qjNz\":[\"Personal Motivations represent what drives your Agent beyond their Bonds. \\nThese can be faith, patriotism, hobbies, or even the love of a pet. While \\npowerful, these motivations aren't as strong as Bonds to other humans.\"],\"9XnWrG\":[\"personal specialty (any)\"],\"IzBQrW\":[\"Personality\"],\"MdYRcY\":[\"Persuade\"],\"XdLRtZ\":[\"pew-pew\"],\"BggFo8\":[\"Pharmacy\"],\"aUuLB9\":[\"Phsyics, Chemistry, Biology, etc\"],\"JRwb4O\":[\"Physician\"],\"4+u/ep\":[\"Physics, Chemistry, Biology, etc\"],\"BVcN3k\":[\"Pilot\"],\"+R3h1p\":[\"Pilot (Aircraft)\"],\"YCHZ9w\":[\"Pleasant\"],\"idLVai\":[\"Pliable\"],\"8TEhqN\":[\"Point Buy\"],\"COUcFd\":[\"Points\"],\"BK4VZy\":[\"Police Officer\"],\"JoUIEx\":[\"Polymath\"],\"NHfDC5\":[\"Power\"],\"n+1nKb\":[\"Preset Professional Skills\"],\"F6Ibbh\":[\"Profession & Skills\"],\"pNZe3j\":[\"Profession Details\"],\"EiRYvW\":[\"Professional Background\"],\"fFTzL+\":[\"Program Manager\"],\"yGzhm3\":[\"Psychological Data\"],\"45eqP/\":[\"Psychotherapy\"],\"VfxeQo\":[\"Punch, kick, grapple, bite, etc\"],\"Xdl7I6\":[\"Radiant\"],\"NrA1Tk\":[\"Raw physical power\"],\"QKcpXG\":[\"Razor-sharp\"],\"37dxPM\":[\"Reads occasionally\"],\"YY/8OS\":[\"Recommended best stats for your chosen profession.\"],\"rYXfOA\":[\"Regional Settings\"],\"XVLIk/\":[\"Removing a bullet, stitching a wound\"],\"YUaj/8\":[\"Repulsive\"],\"/v4BCF\":[\"Reset Bonds\"],\"nV3jkp\":[\"Reset Bonus Points\"],\"oaVDgt\":[\"Reset everything\"],\"/x5uaa\":[\"Reset Personal Details\"],\"MWfUCJ\":[\"Reset Profession\"],\"yKJGwK\":[\"Reset Skill Points\"],\"e69RxD\":[\"Reset Statistics\"],\"hBGyz8\":[\"Reset Stats\"],\"kogYC7\":[\"Resilient\"],\"rbfFz4\":[\"Resolute\"],\"PcdWmG\":[\"Ride\"],\"KZGIQJ\":[\"Robust\"],\"maR6rE\":[\"Roll 4d6, drop lowest Result\"],\"dIRnNt\":[\"Sagely\"],\"46u2kt\":[\"Sailor\"],\"M+hL4r\":[\"Sanity\"],\"5/lolt\":[\"Science\"],\"4/rbeA\":[\"Science (Theoretical Physics)\"],\"RN1nHA\":[\"Scientist\"],\"Eq6YVV\":[\"Score\"],\"A1taO8\":[\"Search\"],\"PHOy1K\":[\"Select Date of Birth\"],\"oE8mTH\":[\"Select Theme\"],\"Tz0i8g\":[\"Settings\"],\"6L2EWT\":[\"SIGINT\"],\"voEyf2\":[\"Signal intelligence, breaking codes\"],\"X1cBOB\":[\"Skill Choices Remaining\"],\"ul6aFX\":[\"Skill Points\"],\"e53RAl\":[\"Skill Points Remaining\"],\"6gjFgQ\":[\"Skill Rating\"],\"professional-guidance.skill-cap-text\":[\"Skills can reach \",[\"DEFAULT_MAX_SKILL_VALUE\"],\"% with point allocation, and hit \",[\"DEFAULT_TOTAL_CAP\"],\"% with bonus points. If your bonus points would exceed \",[\"DEFAULT_TOTAL_CAP\"],\"%, the skill will cap at \",[\"DEFAULT_TOTAL_CAP\"],\"% and anyexcess points will be lost. Skill values can improve beyond \",[\"DEFAULT_TOTAL_CAP\"],\"%\"],\"aQnfn1\":[\"Sneaking around\"],\"HXxT8R\":[\"So much is illegal that there are broad economies of crime. This profile fits a hardened militant or a traditional “black collar” criminal: pimp, burglar, extortionist, or thug. If you want a white-collar criminal, choose Computer Scientist or Business Executive and make very risky decisions.\"],\"bgjaGb\":[\"Soft\"],\"QLiQBs\":[\"Soldier\"],\"KNyHRH\":[\"Something you admire\"],\"6jBuvW\":[\"Something you dislike\"],\"vsXP3+\":[\"Special Operator\"],\"ITqzjt\":[\"Speed and physical agility\"],\"9Cp1yM\":[\"Spouse or ex-spouse\"],\"Pugge/\":[\"Standard Professions\"],\"tXkhj/\":[\"Start\"],\"1ktYIh\":[\"starts at\"],\"24PEWv\":[\"Starts at:\"],\"DrzPEl\":[\"Statistical Data\"],\"2NbyY/\":[\"Statistics\"],\"V2infW\":[\"Statistics Details\"],\"GU/TXk\":[\"Stealth\"],\"KN4iFs\":[\"Stiff\"],\"JR0j+9\":[\"Strength\"],\"6yvfxu\":[\"Stubborn\"],\"n3sBo6\":[\"Study of humans and their cultures\"],\"rKFfgg\":[\"Sturdy\"],\"dXoieq\":[\"Summary\"],\"XYLcNv\":[\"Support\"],\"nAmcJf\":[\"Surgery\"],\"QprZWI\":[\"Surgically Precise\"],\"5ojEcZ\":[\"Survival\"],\"WWQ4IE\":[\"Susceptible\"],\"//qGGh\":[\"Swim\"],\"rk4kEq\":[\"Swole\"],\"niuT0f\":[\"Take a peek behind the curtain. Feel free to:<0><1>report an <2>issue </2></1><3>leave <4>feedback/feature requests</4></3><5>or ⭐️ Star the project to give the author ego a boost.</5></0>\"],\"47lL4r\":[\"Tenacious\"],\"S4gbUS\":[\"The Crucible\"],\"/X90px\":[\"The name of the individual or group your agent is bonded to.\"],\"T4VBOv\":[\"The title of your agent's day-job\"],\"7bPwkx\":[\"Theoretical Physics\"],\"udFMBl\":[\"Thick-skulled\"],\"ndeZXv\":[\"Things Man Was Not Meant to Know\"],\"P2YRxu\":[\"Title:\"],\"aw0XRr\":[\"to purchase Delta Green source materials and learn more.\"],\"72c5Qo\":[\"Total\"],\"u0Vxnx\":[\"Unarmed Combat\"],\"6+uTR4\":[\"Uncanny\"],\"tNJJOF\":[\"Uncoordinated\"],\"K+/GwL\":[\"Understanding and arguing the law\"],\"l2ynLe\":[\"Understanding human behavior\"],\"UyWBj5\":[\"Unhealthy\"],\"Hs/oFo\":[\"Uninspiring\"],\"haSd2S\":[\"Unnatural\"],\"xa8Jrn\":[\"Unremarkable\"],\"BoY0Al\":[\"Unyielding\"],\"tI35bN\":[\"Use All Suggestions\"],\"yjOG8K\":[\"Use Suggested\"],\"COU52/\":[\"Use these prompts to flesh out your character. They are all optional but can help put you in the shoes of your character.\"],\"dsWcdt\":[\"Vigorous\"],\"K2f7tT\":[\"Violence\"],\"ItnLvW\":[\"Visit\"],\"yudl3a\":[\"We're fighting a war against the inevitable. And it's a war we're going to lose. But we're going to win the battles. We're going to make the bastards work for it. We're going to make them take ground inch by inch. And we're going to make them bleed for every inch they take.\"],\"9ls/Vs\":[\"Weak\"],\"k6ZCLg\":[\"Weak-willed\"],\"Vo1ACK\":[\"Weekend Warrior\"],\"/oYuXu\":[\"What happens at the opera stays here. Not that anyone outside of the organization would believe you though.\"],\"P8Xrdt\":[\"What is Delta Green?\"],\"imD/Q0\":[\"Whether aboard a naval vessel or merchant marine ship, you've spent your life on the waves. The sea holds many secrets, and those who spend enough time there learn that some things are better left undisturbed. Something you encountered in those vast waters caught Delta Green's attention—perhaps something that shouldn't have been floating, or a discovery in a depth where nothing should survive.\"],\"DUpGIu\":[\"Whip-smart\"],\"RBd2c3\":[\"Why does Delta Green trust this agent?\"],\"CyfgYa\":[\"Why does this agent serve Delta Green?\"],\"imbYbz\":[\"Will Power\"],\"fDORhQ\":[\"Will, spirit, and mental stability\"],\"kP3ycB\":[\"Willpower\"],\"mYJzKG\":[\"x5\"],\"375wgc\":[\"You expand human knowledge in a field such as biology, physics, or chemistry. When certain forms of knowledge cause insanity and death, it's easy to conclude that some hypotheses should not be tested.\"],\"vTWySo\":[\"You might be an author, an editor, a researcher for a company or any branch of the government, a blog-ger, a TV reporter, or a scholar of rare texts. With the unnatural, you've uncovered the story of a lifetime.\"],\"y87741\":[\"You piece together the truth from fragments of the past - documents, artifacts, and the whispered stories that never made it into textbooks. Your research has led you down paths that mainstream academia refuses to acknowledge, uncovering patterns in human history that suggest something vast and terrible lurking behind civilization's rise and fall. Some archives are sealed for good reason.\"],\"eh3cxY\":[\"You run an organization. Someone has to secure funding, move resources, and make connections, and that's you. You control a budget and are responsible for how your program is maintained and where the money goes. Organizations discover the most startling things in their pursuit of profit or the public good.\"],\"V1GS2V\":[\"You serve and protect. Police officers walk the beat in uniform. Deputy sheriffs answer to an elected law enforcer and have jurisdiction over an entire county. Detectives come in after the fact and put the pieces together.\"],\"eYVzTc\":[\"You study humanity in all its forms - from isolated tribes to urban subcultures. Your work takes you to remote corners of the world where ancient practices persist and forbidden knowledge lingers. Whether you're conducting field interviews in war-torn regions or analyzing cultural artifacts that defy explanation, you've learned that some traditions exist for darker reasons than anyone suspects.\"],\"doadht\":[\"You travel to strange lands, meet interesting people, and try to get along with them. Odds are you work for the State Department, though USAID, the Commercial Service and the Foreign Agriculture Service also have FSOs. Either way, you've had every opportunity to learn exotic and deadly things; the kinds of things that qualify you for Delta Green clearance.\"],\"65+vXE\":[\"You're a member of the United States Army, Marine Corps, or Air Force. You've been trained to fight, and you've been trained to survive. You've seen things that make the average person blanch.\"],\"blWU4y\":[\"Your Agent can have up to five Personal Motivations. Add them during character \\ncreation or as your story develops. Each time your Agent experiences a Breaking \\nPoint, remove one motivation to represent their growing trauma.\"],\"W0MqKV\":[\"Your Agent gains 10% in the Unnatural skill and adds +20% to Occult. Reduce your Agent's SAN by his or her POW. Your Agent gains a new disorder caused by the Unnatural (see page 72). Reset your Agent's Break- ing Point to his or her new SAN minus POW.\"],\"1nTbeT\":[\"Your job oscillates between the tedium of maintaining your gear, exhilaration when the alarm finally comes, and the work of investigating a scene after the smoke has cleared. If you're involved with Delta Green, you clearly stumbled into something worse than a house fire.\"],\"+naB54\":[\"Your tools are a computer and smartphone. You might be moving millions of dollars, or bits of data, or both. Or you might be a prosecutor, a defense attorney, or judge.\"],\"8M2iTK\":[\"Skills can reach \",[\"DEFAULT_MAX_SKILL_VALUE\"],\"% with point allocation, and hit \\n    \",[\"DEFAULT_TOTAL_CAP\"],\"% with bonus points. If your bonus points would exceed \\n    \",[\"DEFAULT_TOTAL_CAP\"],\"%, the skill will cap at \",[\"DEFAULT_TOTAL_CAP\"],\"% and any \\n    excess points will be lost. Skill values can improve beyond \",[\"DEFAULT_TOTAL_CAP\"],\"%\\n    with experience gained during gameplay.\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"stat.reminder\":[[\"reminder\"]],\"stat.label\":[[\"stat\"]],\"stat.intelligence.reminder\":[\"Cunning, logic, and intuition\"],\"stat.constitution.reminder\":[\"Health and hardiness\"],\"stat.charisma.reminder\":[\"Personal and physical appeal\"],\"stat.strength.reminder\":[\"Raw physical power\"],\"stat.dexterity.reminder\":[\"Speed and physical agility\"],\"stat.power.reminder\":[\"Will, spirit, and mental stability\"],\"8VjWl0\":[\"--- Additional Professions ---\"],\"KE5vnC\":[\"--- Standard Professions ---\"],\"7o9kk2\":[\"(Choose type below)\"],\"J/hVSQ\":[[\"0\"]],\"X3V8cR\":[[\"0\"]],\"xYxQCZ\":[[\"0\"],\" \",[\"1\"]],\"Zz4Qvw\":[[\"capitalizedStat\"]],\"w9dLMW\":[[\"label\"]],\"JpzPnP\":[[\"labelText\"]],\"choose-skills.no-choices\":[[\"professionName\"],\" has no flexible skills to choose from.\"],\"leNkYU\":[[\"reminderText\"]],\"5ft2yB\":[[\"skillCapText\"]],\"kDmBhk\":[[\"statKey\"]],\"urTZDx\":[[\"statKey\"]],\"LZJAen\":[\"`$\",[\"0\"],\" has no flexible skills to choose from.`\"],\"9eGA8Z\":[\"010111100001\"],\"Hzaa8H\":[\"1. Last Name, First Name, Middle Initial\"],\"yQJwH3\":[\"10. Physical Description\"],\"kmhPgQ\":[\"11. Bonds\"],\"yGWS1j\":[\"12. Motivations and Mental Disorders\"],\"w/Xy4v\":[\"13. Incidents of San Loss Without Going Insane\"],\"VnPSmP\":[\"2. Profession (Rank if Applicable)\"],\"qVQ1QT\":[\"3. Employer\"],\"nVobO+\":[\"4. Nationality\"],\"icgCRe\":[\"5. Sex\"],\"+18IK1\":[\"6. Age and D.O.B.\"],\"AbZcNQ\":[\"7. Education and Occupational History\"],\"bzqTDJ\":[\"8. Statistics\"],\"W5FXNn\":[\"9. Derived Attributes\"],\"5XYRnE\":[\"A dedicated hobbyist; with a foreign language you can have rudimentary conversations.\"],\"igi2jT\":[\"A lifetime’s pursuit or multiple doctorates.\"],\"uyJsf6\":[\"About\"],\"2zPUa5\":[\"About the Author\"],\"UNDRet\":[\"About the Game\"],\"WcvywZ\":[\"Accounting\"],\"F6pfE9\":[\"Active\"],\"zq0AxB\":[\"Active Couch Potato\"],\"yGwMqg\":[\"adapted\"],\"0kRNUo\":[\"Add +10% to your Agent's Occult and +10% to any five skills other than Unnatural. This can bring no skill higher than 90%. Reduce your Agent's SAN by 5. Remove one Bond.\"],\"Xi6Is8\":[\"Add +10% to your Agent's Occult skill. Reduce SAN by 5. Subtract 3 from your Agent's CHA and each Bond. Your Agent is adapted to violence (see page 73).\"],\"EHRBpm\":[\"Add +10% to your Agent's Occult skill. Reduce SAN by 5. Subtract 3 from your Agent's POW. Your Agent is adapted to helplessness (see page 73).\"],\"custom-skill-form.bonus-points-reminder\":[\"Adds \",[\"DEFAULT_BONUS_VALUE\"],\" to the skill total for each bonus point allocated (capped at \",[\"DEFAULT_TOTAL_CAP\"],\")\"],\"7TlSPN\":[\"Adroit\"],\"zVgowR\":[\"Affable\"],\"CmUtio\":[\"Agile\"],\"gXy38H\":[\"Air or sea, commercial or military, your duty is to keep your passengers alive and craft intact. This can lead to hard choices when your passengers put the vehicle in danger. Or are you a drone operator, flying a Predator from a thousand miles away? Either way, what op brought you to the attention of Delta Green?\"],\"Z+c4we\":[\"Aircraft\"],\"rPKlAj\":[\"Alertness\"],\"EbuTi3\":[\"Alias\"],\"JIeg5e\":[\"All Thumbs\"],\"6zd6CL\":[\"Alluring\"],\"UCIiuP\":[\"Analysing thought\"],\"WbrlFU\":[\"Andrew Sobiesiak is a Software Engineer that has played TTRPGs online weekly since 2017.<0/> He would make a terrible field agent but is prepared to fight the sisyphean cosmic war. He Lives in <1>Seattle, WA</1> with his <2>wife, 5 cats, and a dog</2>\"],\"2qgwCU\":[\"Anthropologist\"],\"juHsSO\":[\"Anthropology\"],\"aAIQg2\":[\"Appearance\"],\"f+WR1d\":[\"Applicable Skill Sets\"],\"FGdRNI\":[\"Apply preset skills and choose additional skills for your character.\"],\"xLXsjU\":[\"Arc Dream Publishing is a small-press publisher that produces high-quality games, including roleplaying games, a horror gaming magazine, and card games. Some of their popular works include Delta Green, Swords & Sorceries, and The Black Company RPG.\"],\"vJo5g6\":[\"Archeology\"],\"+QARA4\":[\"Art\"],\"Zvntde\":[\"Art (Painting)\"],\"ziY9j1\":[\"Artillery\"],\"3Vj2Gv\":[\"As a member of the United States Marine Corps, you are part of America's elite expeditionary force in readiness. First to fight, you've been trained to operate in any environment, from urban warfare to amphibious operations. Your intense training and esprit de corps set you apart. Whatever caught Delta Green's attention, it wasn't your first brush with hell.\"],\"EOS+bp\":[\"As part of a force like the U.S. Army Rangers, you volunteered for a more difficult path than other soldiers. You've spent years in the most grueling training on the planet, and now serve on the most dangerous missions around.\"],\"vS6YOc\":[\"Athletic\"],\"JGpQgX\":[\"Athletics\"],\"BO9EJL\":[\"Avoiding getting hit\"],\"GDkUlj\":[\"Awkward\"],\"ZuyHGl\":[\"Base Statistics\"],\"o5DA10\":[\"Beliefs\"],\"Bj0YCx\":[\"Bland\"],\"pFWzUl\":[\"Bond description\"],\"joDnA9\":[\"Bond Detail\"],\"Qp5ZKm\":[\"Bond Name\"],\"004WwA\":[\"Bond Type\"],\"MTMViL\":[\"Bonds\"],\"ooT9hB\":[\"Bonds Details\"],\"YfLs0q\":[\"Bonds represent meaningful relationships your agent has with non-player characters.\"],\"7zP6Gy\":[\"Bonds represent your Agent's relationships with other people.\\n                    Each bond starts with a score equal to your Agent's Charisma (\",[\"0\"],\").\\n                    The score of a bond can improve if it is nurtured between operations. A bond can never have a score higher than your Agent's Charisma.\\n                    Any time your Agent's Charisma drops, each bond score drops by the same amount.\\n                    Bonds can be defined as a single person or a group of people. For example, an agent might have bonds with their spouse, their business partner, and their bowling league.\\n                    A Bond must be a real person or small group of people who are alive and can be interacted with.\"],\"PqoQ2W\":[\"Bonus\"],\"Mxgo1h\":[\"Bonus Points\"],\"jw0SVP\":[\"Bonus Points Remaining\"],\"MIHH6C\":[\"Breaking Point\"],\"jEvJsi\":[\"Bright\"],\"GtWnxo\":[\"Brilliant\"],\"Nr0SIL\":[\"Bureaucracy\"],\"CBGYlg\":[\"Business Math\"],\"6rScUj\":[\"Camping, tracking, improvising in nature\"],\"a46V7b\":[\"Captivating\"],\"aw5Ht8\":[\"Captivity or Imprisonment\"],\"5NsOuM\":[\"Cat-like\"],\"hVHY2j\":[\"Charisma\"],\"gLHd7f\":[\"Check a Bond’s box when projecting sanity damage\"],\"DK/qpg\":[\"Check a skill when you fail a skill check. After a session, add 1d4 to each checked skill and erase the check.\"],\"nUIcQr\":[\"Choose \",[\"0\"],\" Additional Skills\"],\"zikS7b\":[\"Choose a bonus skill package\"],\"k/xMy9\":[\"Choose a package to automatically apply bonus points, or select 'No Package' to manually assign your bonus points\"],\"CGI4Ui\":[\"Church or support group\"],\"xCJdfg\":[\"Clear\"],\"Jz0/oR\":[\"Clear Descriptors\"],\"LndrJY\":[\"Clear Selected Skills\"],\"Ql//nl\":[\"Clear Templates\"],\"OxOc2C\":[\"Clumsy\"],\"RLWPXk\":[\"Colleagues in an intense job\"],\"KLEQcs\":[\"College major or years of experience; with 50% in a foreign language you have native fluency.\"],\"rIdFzg\":[\"College minor or basic training.\"],\"yKoSLY\":[\"Compelling\"],\"kHbyNa\":[\"Computer Science\"],\"AU0Kdy\":[\"Computers and machinery are the backbone of modern industry. You are a craftsman with data or machinery, possibly for the government and most definitely for profit. However you use your skills, the overlap between information technology and awareness of the unnatural could make this the most dangerous job on the planet.\"],\"fOaHdd\":[\"Constitution\"],\"uhhJuh\":[\"Controlled explosions\"],\"IbtVsv\":[\"Convincing people\"],\"3zF2gs\":[\"Couch Potato\"],\"gtWP1l\":[\"Could take the stairs\"],\"A8wadO\":[\"Coworker or partner\"],\"UTfQwZ\":[\"Craft\"],\"470Fk4\":[\"Craft (Macrame)\"],\"XjzLm3\":[\"Creative pursuits\"],\"PMunmF\":[\"Criminal\"],\"oWJvZ7\":[\"Criminology\"],\"VoIJ7Q\":[\"CSI\"],\"5ALalD\":[\"Cult shit\"],\"d/rSvQ\":[\"Cunning, logic, and intuition\"],\"Hp1l6f\":[\"Current\"],\"zIo0DE\":[\"Custom Professions\"],\"7/D6Vk\":[\"Dabbler.\"],\"iX6K3g\":[\"Damaged Veteran Templates\"],\"7uHWR+\":[\"Date of Birth\"],\"onAw8T\":[\"Decades of experience, or a graduate or doctoral degree.\"],\"tffdBt\":[\"Declassified Record\"],\"L3rQ+j\":[\"Deft\"],\"VVywaC\":[\"Delicate\"],\"PgSdTp\":[\"Demolitions\"],\"ezghSE\":[\"Derived Attributes\"],\"Nu4oKW\":[\"Description\"],\"s3Uzfx\":[\"Determined\"],\"08z1KE\":[\"Dexterity\"],\"u9l1aF\":[\"Dice\"],\"DwvMe2\":[\"Dim\"],\"BJ6YQv\":[\"Disguise\"],\"uPWSEh\":[\"Distinguishing Features\"],\"9/Nzrk\":[\"Do you close your eyes to what you’ve seen and go back to sleep? Or do you come with this psycho burnout and do the impossible against the unbelievable and keep the future at bay for another day?\"],\"difbU5\":[\"Do you like drugs?\"],\"DM17Gg\":[\"Doctors are often the first to uncover signs of an unnatural incursion, and the most valuable investigators of its disastrous effects on humanity.\"],\"p+tvhp\":[\"Dodge\"],\"OUmX26\":[\"Dossier\"],\"Q2ObDC\":[\"Drive\"],\"1afgyS\":[\"Dull\"],\"aqxYLv\":[\"Education\"],\"n9u/m0\":[\"Employer\"],\"pbUVnA\":[\"Engaging\"],\"iy5fSS\":[\"Engineer\"],\"RI/fzG\":[\"Enter a Subtype\"],\"HmI5oK\":[\"Example\"],\"O4wxl8\":[\"Export as PDF\"],\"S8GwkS\":[\"Extreme Violence\"],\"F7dRcr\":[\"Fair\"],\"+UN+Ez\":[\"Fatigued\"],\"y+uhgI\":[\"Federal Agent\"],\"hrDJ2/\":[\"Feeble\"],\"oCrptp\":[\"Finding the path\"],\"qUl/bH\":[\"Finding things\"],\"GnG+vN\":[\"Firearms\"],\"/OsYEc\":[\"Firefighter\"],\"RtEHo7\":[\"Firm\"],\"11+tdk\":[\"First Aid\"],\"kODvZJ\":[\"First Name\"],\"ZX5IeB\":[\"First Night at the Opera?\"],\"u4iXQi\":[\"Fit\"],\"wMLcWn\":[\"Flexible in principle\"],\"yaw+7x\":[\"Font Settings\"],\"kXTwNy\":[\"Foreign Languages\"],\"aFFQEb\":[\"Foreign Languages (French)\"],\"a9k1oC\":[\"Foreign Service Officer\"],\"z7q/DL\":[\"Forensics\"],\"K9cGgQ\":[\"Forgettable\"],\"0BlDs+\":[\"Forklifts, cranes, excavators, etc\"],\"bdeAb2\":[\"Formidable\"],\"noN438\":[\"Fortified\"],\"1Yrw83\":[\"Found on page \",[\"pageNumber\"]],\"R2sFu3\":[\"Fragile\"],\"ZQfuYY\":[\"Frail\"],\"nLC6tu\":[\"French\"],\"yTcD6q\":[\"Fumbling\"],\"ZJ1pST\":[\"Gene Parmesan, he's the best\"],\"7o4S1E\":[\"Genius\"],\"+/PqzT\":[\"Graceless\"],\"47ofD/\":[\"Greased Lightning\"],\"fcFvf1\":[\"Greasing the wheels of government\"],\"L8fEEm\":[\"Group\"],\"J36o9O\":[\"Handy\"],\"fjd2D3\":[\"Hard Experience\"],\"bzM4ak\":[\"Health and hardiness\"],\"yGS9cI\":[\"Healthy\"],\"UjSOv0\":[\"Hearty\"],\"NxZ7K+\":[\"Heavy Machinery\"],\"Nw0C1v\":[\"Heavy Weapons\"],\"0ZBnED\":[\"Helplessness\"],\"ORe4Lw\":[\"Historian\"],\"0caMy7\":[\"History\"],\"Bwb/M2\":[\"Hit Points\"],\"1+WSY1\":[\"Hobbies\"],\"i0qMbr\":[\"Home\"],\"UU7+rG\":[\"Horses and such\"],\"21qVcE\":[\"Huge\"],\"Po1nwb\":[\"HUMINT\"],\"4bNpE0\":[\"If the intelligence community was a family, Delta Green would be the uncle nobody talks about.<0/>Delta Green is a Table Top Role Playing Game that uses a D100 system. Handlers run operations for their agents, other players, where a grey government operation works to protect the world from unkown cosmic horrors.\"],\"dbLUe7\":[\"If you haven't studied it you're doomed to repeat it\"],\"p64A7d\":[\"If you want to support this work, you can buy some Delta Green content through affiliate links.\"],\"NNy6W8\":[\"Impulsive\"],\"b23dA7\":[\"In the FBI, NSA and CIA, there are those who gather information and those who decide what it means. You take information from disparate sources—newspapers, websites, informants, ELINT, and the assets developed by Case Officers—and figure out what it means. In short, your job is the piecing together of unrelated knowledge, a dangerous endeavor in the world of Delta Green.\"],\"UFi3x2\":[\"Incurious\"],\"fByw/g\":[\"Individual\"],\"+RRbXc\":[\"Indomitable\"],\"aemL3D\":[\"Infirmed\"],\"C5kxqk\":[\"Inner Adam Savage\"],\"s2/OPX\":[\"Insipid\"],\"Gzw2pq\":[\"Intelligence\"],\"Mf5ErF\":[\"Intelligence Analyst\"],\"3FDXvB\":[\"Intelligence Case Officer\"],\"VGX26E\":[\"Irrational\"],\"Jp7JUr\":[\"It's a jeep thing, you wouldn't understand\"],\"mEzWf0\":[\"Jacked\"],\"tcBJo7\":[\"Jumping, ducking, running, climbing, etc\"],\"yoMIHR\":[\"Jurassic Park\"],\"sNoMVY\":[\"Key Stats\"],\"zC0r/K\":[\"Knives, hatchets, swords, etc\"],\"87a/t/\":[\"Label\"],\"ok/7qO\":[\"Land\"],\"UXBCwc\":[\"Last Name\"],\"iqhOTb\":[\"Law\"],\"whBNst\":[\"Lawyer\"],\"XHdFzA\":[\"List character motivations and any mental disorders...\"],\"AqDRRD\":[\"LMGs I think\"],\"1z0Tzf\":[\"Macrame\"],\"kT7U6s\":[\"Magnetic\"],\"oLEThm\":[\"Making things explode from far away\"],\"PONy2U\":[\"Manual Input\"],\"LlSUa5\":[\"Many Delta Green Agents are federal law enforcement officers, mostly from the FBI. Delta Green decided long ago that federal agents have the optimum balance of skills and mental stability needed to confront the unnatural.\"],\"PAXVdT\":[\"Marine\"],\"CK1KXz\":[\"Max\"],\"spK3cv\":[\"Media Specialist\"],\"zeVoXl\":[\"Medical practice\"],\"hS5cj7\":[\"Medical professionals are on the front line when awful things happen. Is that what brought you to the group's attention?\"],\"6Kw9Gv\":[\"Medicine\"],\"F5rArt\":[\"Mediocre\"],\"syonIH\":[\"Meet me at <0>The Crucible</0> and tell me a little bit about yourself, agent. What's your deal?\"],\"qx+7SQ\":[\"Melee Weapons\"],\"e3Oj4X\":[\"Mesmerizing\"],\"2NOxz2\":[\"Middle Initial\"],\"dImRx9\":[\"Military Science\"],\"YOtRZ0\":[\"Military Science (Land)\"],\"hNi0vh\":[\"Military tactics\"],\"lJ71dZ\":[\"Minor urgent medical care\"],\"dOLpsA\":[\"Motivations\"],\"aijpeY\":[\"Moving oneself through water\"],\"e9CT52\":[\"Muddying the waters of a crime-scene\"],\"av28MK\":[\"Muscular\"],\"2YKBHm\":[\"Nationality\"],\"dbDjdV\":[\"Navigate\"],\"mqUHTq\":[\"Nigh Immortal\"],\"Ysyt+a\":[\"Nimble\"],\"kSzR8n\":[\"No Package (Manual Bonus Points)\"],\"u3hw5g\":[\"Note additional details about the bond\"],\"XOGTWH\":[\"Noticing things\"],\"9OGltc\":[\"Number of social connections available to your character.\"],\"tEPJOu\":[\"Nurse\"],\"xhyJlX\":[\"Obsessions\"],\"7JOZLS\":[\"Occult\"],\"i0qI4Z\":[\"Opearting a motor-vehicle\"],\"ZI3xDG\":[\"Operating flying vehicles\"],\"0zpgxV\":[\"Options\"],\"DW8nOh\":[\"Oracular\"],\"08AP9H\":[\"Painting\"],\"pYwzJ9\":[\"Parles-vous Français?\"],\"G/Cl3b\":[\"Pays for Mensa\"],\"Xvtlo9\":[\"Perserverant\"],\"iLp4WI\":[\"Personal and physical appeal\"],\"DhAC6+\":[\"Personal Data\"],\"sJYPeP\":[\"Personal Details\"],\"8peLId\":[\"Personal Motivation \",[\"0\"]],\"VpAfxU\":[\"Personal Motivations\"],\"6dVMln\":[\"Personal Motivations represent what drives your Agent beyond their Bonds. \\n    These can be faith, patriotism, hobbies, or even the love of a pet. While \\n    powerful, these motivations aren't as strong as Bonds to other humans.\"],\"39qjNz\":[\"Personal Motivations represent what drives your Agent beyond their Bonds. \\nThese can be faith, patriotism, hobbies, or even the love of a pet. While \\npowerful, these motivations aren't as strong as Bonds to other humans.\"],\"9XnWrG\":[\"personal specialty (any)\"],\"IzBQrW\":[\"Personality\"],\"MdYRcY\":[\"Persuade\"],\"XdLRtZ\":[\"pew-pew\"],\"BggFo8\":[\"Pharmacy\"],\"aUuLB9\":[\"Phsyics, Chemistry, Biology, etc\"],\"JRwb4O\":[\"Physician\"],\"4+u/ep\":[\"Physics, Chemistry, Biology, etc\"],\"BVcN3k\":[\"Pilot\"],\"+R3h1p\":[\"Pilot (Aircraft)\"],\"YCHZ9w\":[\"Pleasant\"],\"idLVai\":[\"Pliable\"],\"8TEhqN\":[\"Point Buy\"],\"COUcFd\":[\"Points\"],\"BK4VZy\":[\"Police Officer\"],\"JoUIEx\":[\"Polymath\"],\"NHfDC5\":[\"Power\"],\"n+1nKb\":[\"Preset Professional Skills\"],\"F6Ibbh\":[\"Profession & Skills\"],\"pNZe3j\":[\"Profession Details\"],\"EiRYvW\":[\"Professional Background\"],\"fFTzL+\":[\"Program Manager\"],\"yGzhm3\":[\"Psychological Data\"],\"45eqP/\":[\"Psychotherapy\"],\"VfxeQo\":[\"Punch, kick, grapple, bite, etc\"],\"Xdl7I6\":[\"Radiant\"],\"NrA1Tk\":[\"Raw physical power\"],\"QKcpXG\":[\"Razor-sharp\"],\"37dxPM\":[\"Reads occasionally\"],\"YY/8OS\":[\"Recommended best stats for your chosen profession.\"],\"rYXfOA\":[\"Regional Settings\"],\"XVLIk/\":[\"Removing a bullet, stitching a wound\"],\"YUaj/8\":[\"Repulsive\"],\"/v4BCF\":[\"Reset Bonds\"],\"nV3jkp\":[\"Reset Bonus Points\"],\"oaVDgt\":[\"Reset everything\"],\"/x5uaa\":[\"Reset Personal Details\"],\"MWfUCJ\":[\"Reset Profession\"],\"yKJGwK\":[\"Reset Skill Points\"],\"e69RxD\":[\"Reset Statistics\"],\"hBGyz8\":[\"Reset Stats\"],\"kogYC7\":[\"Resilient\"],\"rbfFz4\":[\"Resolute\"],\"PcdWmG\":[\"Ride\"],\"KZGIQJ\":[\"Robust\"],\"maR6rE\":[\"Roll 4d6, drop lowest Result\"],\"dIRnNt\":[\"Sagely\"],\"46u2kt\":[\"Sailor\"],\"M+hL4r\":[\"Sanity\"],\"5/lolt\":[\"Science\"],\"4/rbeA\":[\"Science (Theoretical Physics)\"],\"RN1nHA\":[\"Scientist\"],\"Eq6YVV\":[\"Score\"],\"A1taO8\":[\"Search\"],\"PHOy1K\":[\"Select Date of Birth\"],\"oE8mTH\":[\"Select Theme\"],\"Tz0i8g\":[\"Settings\"],\"6L2EWT\":[\"SIGINT\"],\"voEyf2\":[\"Signal intelligence, breaking codes\"],\"X1cBOB\":[\"Skill Choices Remaining\"],\"ul6aFX\":[\"Skill Points\"],\"e53RAl\":[\"Skill Points Remaining\"],\"6gjFgQ\":[\"Skill Rating\"],\"professional-guidance.skill-cap-text\":[\"Skills can reach \",[\"DEFAULT_MAX_SKILL_VALUE\"],\"% with point allocation, and hit \",[\"DEFAULT_TOTAL_CAP\"],\"% with bonus points. If your bonus points would exceed \",[\"DEFAULT_TOTAL_CAP\"],\"%, the skill will cap at \",[\"DEFAULT_TOTAL_CAP\"],\"% and anyexcess points will be lost. Skill values can improve beyond \",[\"DEFAULT_TOTAL_CAP\"],\"%\"],\"aQnfn1\":[\"Sneaking around\"],\"HXxT8R\":[\"So much is illegal that there are broad economies of crime. This profile fits a hardened militant or a traditional “black collar” criminal: pimp, burglar, extortionist, or thug. If you want a white-collar criminal, choose Computer Scientist or Business Executive and make very risky decisions.\"],\"bgjaGb\":[\"Soft\"],\"QLiQBs\":[\"Soldier\"],\"KNyHRH\":[\"Something you admire\"],\"6jBuvW\":[\"Something you dislike\"],\"vsXP3+\":[\"Special Operator\"],\"ITqzjt\":[\"Speed and physical agility\"],\"9Cp1yM\":[\"Spouse or ex-spouse\"],\"Pugge/\":[\"Standard Professions\"],\"tXkhj/\":[\"Start\"],\"1ktYIh\":[\"starts at\"],\"24PEWv\":[\"Starts at:\"],\"DrzPEl\":[\"Statistical Data\"],\"2NbyY/\":[\"Statistics\"],\"V2infW\":[\"Statistics Details\"],\"GU/TXk\":[\"Stealth\"],\"KN4iFs\":[\"Stiff\"],\"JR0j+9\":[\"Strength\"],\"6yvfxu\":[\"Stubborn\"],\"n3sBo6\":[\"Study of humans and their cultures\"],\"rKFfgg\":[\"Sturdy\"],\"dXoieq\":[\"Summary\"],\"XYLcNv\":[\"Support\"],\"nAmcJf\":[\"Surgery\"],\"QprZWI\":[\"Surgically Precise\"],\"5ojEcZ\":[\"Survival\"],\"WWQ4IE\":[\"Susceptible\"],\"//qGGh\":[\"Swim\"],\"rk4kEq\":[\"Swole\"],\"niuT0f\":[\"Take a peek behind the curtain. Feel free to:<0><1>report an <2>issue </2></1><3>leave <4>feedback/feature requests</4></3><5>or ⭐️ Star the project to give the author ego a boost.</5></0>\"],\"47lL4r\":[\"Tenacious\"],\"S4gbUS\":[\"The Crucible\"],\"/X90px\":[\"The name of the individual or group your agent is bonded to.\"],\"T4VBOv\":[\"The title of your agent's day-job\"],\"7bPwkx\":[\"Theoretical Physics\"],\"udFMBl\":[\"Thick-skulled\"],\"ndeZXv\":[\"Things Man Was Not Meant to Know\"],\"P2YRxu\":[\"Title:\"],\"aw0XRr\":[\"to purchase Delta Green source materials and learn more.\"],\"72c5Qo\":[\"Total\"],\"u0Vxnx\":[\"Unarmed Combat\"],\"6+uTR4\":[\"Uncanny\"],\"tNJJOF\":[\"Uncoordinated\"],\"K+/GwL\":[\"Understanding and arguing the law\"],\"l2ynLe\":[\"Understanding human behavior\"],\"UyWBj5\":[\"Unhealthy\"],\"Hs/oFo\":[\"Uninspiring\"],\"haSd2S\":[\"Unnatural\"],\"xa8Jrn\":[\"Unremarkable\"],\"BoY0Al\":[\"Unyielding\"],\"tI35bN\":[\"Use All Suggestions\"],\"yjOG8K\":[\"Use Suggested\"],\"COU52/\":[\"Use these prompts to flesh out your character. They are all optional but can help put you in the shoes of your character.\"],\"dsWcdt\":[\"Vigorous\"],\"K2f7tT\":[\"Violence\"],\"ItnLvW\":[\"Visit\"],\"yudl3a\":[\"We're fighting a war against the inevitable. And it's a war we're going to lose. But we're going to win the battles. We're going to make the bastards work for it. We're going to make them take ground inch by inch. And we're going to make them bleed for every inch they take.\"],\"9ls/Vs\":[\"Weak\"],\"k6ZCLg\":[\"Weak-willed\"],\"Vo1ACK\":[\"Weekend Warrior\"],\"/oYuXu\":[\"What happens at the opera stays here. Not that anyone outside of the organization would believe you though.\"],\"P8Xrdt\":[\"What is Delta Green?\"],\"imD/Q0\":[\"Whether aboard a naval vessel or merchant marine ship, you've spent your life on the waves. The sea holds many secrets, and those who spend enough time there learn that some things are better left undisturbed. Something you encountered in those vast waters caught Delta Green's attention—perhaps something that shouldn't have been floating, or a discovery in a depth where nothing should survive.\"],\"DUpGIu\":[\"Whip-smart\"],\"RBd2c3\":[\"Why does Delta Green trust this agent?\"],\"CyfgYa\":[\"Why does this agent serve Delta Green?\"],\"imbYbz\":[\"Will Power\"],\"fDORhQ\":[\"Will, spirit, and mental stability\"],\"kP3ycB\":[\"Willpower\"],\"mYJzKG\":[\"x5\"],\"375wgc\":[\"You expand human knowledge in a field such as biology, physics, or chemistry. When certain forms of knowledge cause insanity and death, it's easy to conclude that some hypotheses should not be tested.\"],\"vTWySo\":[\"You might be an author, an editor, a researcher for a company or any branch of the government, a blog-ger, a TV reporter, or a scholar of rare texts. With the unnatural, you've uncovered the story of a lifetime.\"],\"y87741\":[\"You piece together the truth from fragments of the past - documents, artifacts, and the whispered stories that never made it into textbooks. Your research has led you down paths that mainstream academia refuses to acknowledge, uncovering patterns in human history that suggest something vast and terrible lurking behind civilization's rise and fall. Some archives are sealed for good reason.\"],\"eh3cxY\":[\"You run an organization. Someone has to secure funding, move resources, and make connections, and that's you. You control a budget and are responsible for how your program is maintained and where the money goes. Organizations discover the most startling things in their pursuit of profit or the public good.\"],\"V1GS2V\":[\"You serve and protect. Police officers walk the beat in uniform. Deputy sheriffs answer to an elected law enforcer and have jurisdiction over an entire county. Detectives come in after the fact and put the pieces together.\"],\"eYVzTc\":[\"You study humanity in all its forms - from isolated tribes to urban subcultures. Your work takes you to remote corners of the world where ancient practices persist and forbidden knowledge lingers. Whether you're conducting field interviews in war-torn regions or analyzing cultural artifacts that defy explanation, you've learned that some traditions exist for darker reasons than anyone suspects.\"],\"doadht\":[\"You travel to strange lands, meet interesting people, and try to get along with them. Odds are you work for the State Department, though USAID, the Commercial Service and the Foreign Agriculture Service also have FSOs. Either way, you've had every opportunity to learn exotic and deadly things; the kinds of things that qualify you for Delta Green clearance.\"],\"65+vXE\":[\"You're a member of the United States Army, Marine Corps, or Air Force. You've been trained to fight, and you've been trained to survive. You've seen things that make the average person blanch.\"],\"2At5A2\":[\"Your Agent can have up to five Personal Motivations. Add them during character \\n    creation or as your story develops. Each time your Agent experiences a Breaking \\n    Point, remove one motivation to represent their growing trauma.\"],\"blWU4y\":[\"Your Agent can have up to five Personal Motivations. Add them during character \\ncreation or as your story develops. Each time your Agent experiences a Breaking \\nPoint, remove one motivation to represent their growing trauma.\"],\"W0MqKV\":[\"Your Agent gains 10% in the Unnatural skill and adds +20% to Occult. Reduce your Agent's SAN by his or her POW. Your Agent gains a new disorder caused by the Unnatural (see page 72). Reset your Agent's Break- ing Point to his or her new SAN minus POW.\"],\"1nTbeT\":[\"Your job oscillates between the tedium of maintaining your gear, exhilaration when the alarm finally comes, and the work of investigating a scene after the smoke has cleared. If you're involved with Delta Green, you clearly stumbled into something worse than a house fire.\"],\"+naB54\":[\"Your tools are a computer and smartphone. You might be moving millions of dollars, or bits of data, or both. Or you might be a prosecutor, a defense attorney, or judge.\"],\"8M2iTK\":[\"Skills can reach \",[\"DEFAULT_MAX_SKILL_VALUE\"],\"% with point allocation, and hit \\n    \",[\"DEFAULT_TOTAL_CAP\"],\"% with bonus points. If your bonus points would exceed \\n    \",[\"DEFAULT_TOTAL_CAP\"],\"%, the skill will cap at \",[\"DEFAULT_TOTAL_CAP\"],\"% and any \\n    excess points will be lost. Skill values can improve beyond \",[\"DEFAULT_TOTAL_CAP\"],\"%\\n    with experience gained during gameplay.\"]}")as Messages;
