+++ src/locales/fr/messages.ts
@@ -1,1 +1,1 @@
-/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"stat.reminder\":[[\"reminder\"]],\"stat.label\":[[\"stat\"]],\"stat.intelligence.reminder\":[\"Cunning, logic, and intuition\"],\"stat.constitution.reminder\":[\"Health and hardiness\"],\"stat.charisma.reminder\":[\"Personal and physical appeal\"],\"stat.strength.reminder\":[\"Raw physical power\"],\"stat.dexterity.reminder\":[\"Speed and physical agility\"],\"stat.power.reminder\":[\"Will, spirit, and mental stability\"],\"8VjWl0\":[\"--- Professions supplémentaires ---\"],\"KE5vnC\":[\"--- Professions standard ---\"],\"7o9kk2\":[\"(Choisir un type ci-dessous)\"],\"J/hVSQ\":[[\"0\"]],\"X3V8cR\":[[\"0\"]],\"xYxQCZ\":[[\"0\"],\" \",[\"1\"]],\"Zz4Qvw\":[[\"capitalizedStat\"]],\"w9dLMW\":[[\"label\"]],\"JpzPnP\":[[\"labelText\"]],\"choose-skills.no-choices\":[[\"professionName\"],\" n'a pas de compétences flexibles à choisir.\"],\"leNkYU\":[[\"reminderText\"]],\"5ft2yB\":[[\"skillCapText\"]],\"kDmBhk\":[[\"statKey\"]],\"urTZDx\":[[\"statKey\"]],\"LZJAen\":[\"`$\",[\"0\"],\" has no flexible skills to choose from.`\"],\"9eGA8Z\":[\"010111100001\"],\"Hzaa8H\":[\"1. Nom, Prénom, Initiale du milieu\"],\"yQJwH3\":[\"10. Description physique\"],\"kmhPgQ\":[\"11. Liens\"],\"yGWS1j\":[\"12. Motivations et troubles mentaux\"],\"w/Xy4v\":[\"13. Incidents de perte de SAN sans devenir fou\"],\"VnPSmP\":[\"2. Profession (Grade si applicable)\"],\"qVQ1QT\":[\"3. Employeur\"],\"nVobO+\":[\"4. Nationalité\"],\"icgCRe\":[\"5. Sexe\"],\"+18IK1\":[\"6. Âge et date de naissance\"],\"AbZcNQ\":[\"7. Formation et histoire professionnelle\"],\"bzqTDJ\":[\"8. Statistiques\"],\"W5FXNn\":[\"9. Attributs dérivés\"],\"5XYRnE\":[\"Un amateur dévoué; avec une langue étrangère, vous pouvez avoir des conversations rudimentaires.\"],\"igi2jT\":[\"Une poursuite de toute une vie ou plusieurs doctorats.\"],\"uyJsf6\":[\"À propos\"],\"2zPUa5\":[\"À propos de l'auteur\"],\"UNDRet\":[\"À propos du jeu\"],\"WcvywZ\":[\"Comptabilité\"],\"F6pfE9\":[\"Actif\"],\"zq0AxB\":[\"Actif couch potato\"],\"yGwMqg\":[\"adapté\"],\"0kRNUo\":[\"Ajoutez +10% à la compétence Non naturel de votre Agent et +10% à cinq compétences autres que Non naturel. Cela ne peut pas amener une compétence à plus de 90%. Réduisez la SAN de votre Agent de 5. Retirez un lien.\"],\"Xi6Is8\":[\"Ajoutez +10% à la compétence Non naturel de votre Agent. Réduisez la SAN de 5. Retirez 3 de la CHA de votre Agent et de chaque lien. Votre Agent est adapté à la violence (voir page 73).\"],\"EHRBpm\":[\"Ajoutez +10% à la compétence Non naturel de votre Agent. Réduisez la SAN de 5. Retirez 3 de la POW de votre Agent. Votre Agent est adapté à l'impuissance (voir page 73).\"],\"custom-skill-form.bonus-points-reminder\":[\"Ajoute \",[\"DEFAULT_BONUS_VALUE\"],\" à la compétence totale pour chaque point bonus alloué (limité à \",[\"DEFAULT_TOTAL_CAP\"],\")\"],\"7TlSPN\":[\"Adroit\"],\"zVgowR\":[\"Affable\"],\"CmUtio\":[\"Agile\"],\"gXy38H\":[\"Air ou mer, commercial ou militaire, votre devoir est de maintenir vos passagers en vie et votre bateau intact. Cela peut mener à des choix difficiles lorsque vos passagers mettent le véhicule en danger. Ou êtes-vous un opérateur de drone, pilotant un Predator à mille miles de distance ? Quoi qu'il en soit, quelle opération vous a fait remarquer à Delta Green ?\"],\"Z+c4we\":[\"Aéronef\"],\"rPKlAj\":[\"Alerte\"],\"EbuTi3\":[\"Pseudonyme\"],\"JIeg5e\":[\"Tous les doigts\"],\"6zd6CL\":[\"Attirant\"],\"UCIiuP\":[\"Analyse du raisonnement\"],\"WbrlFU\":[\"Andrew Sobiesiak est un ingénieur logiciel qui joue aux JDR sur table en ligne chaque semaine depuis 2017.<0/> Il ferait un terrible agent de terrain mais est prêt à combattre la guerre cosmique sisyphéenne. Il vit à <1>Seattle, WA</1> avec sa <2>femme, 5 chats et un chien</2>\"],\"2qgwCU\":[\"Anthropologue\"],\"juHsSO\":[\"Anthropologie\"],\"aAIQg2\":[\"Apparence\"],\"f+WR1d\":[\"Compétences applicables\"],\"FGdRNI\":[\"Appliquer les compétences prédéfinies et choisir des compétences supplémentaires pour votre personnage.\"],\"xLXsjU\":[\"Arc Dream Publishing est un petit éditeur qui produit des jeux de haute qualité, notamment des jeux de rôle, un magazine de jeux d'horreur et des jeux de cartes. Parmi leurs œuvres populaires figurent Delta Green, Swords & Sorceries et The Black Company RPG.\"],\"vJo5g6\":[\"Archéologie\"],\"+QARA4\":[\"Art\"],\"Zvntde\":[\"Art (Peinture)\"],\"ziY9j1\":[\"Artillerie\"],\"3Vj2Gv\":[\"En tant que membre des Marines des États-Unis, vous faites partie de la force expéditionnaire élite des États-Unis en état de préparation. Premier à combattre, vous avez été formé à opérer dans n'importe quelle environnement, de la guerre urbaine à des opérations amphibies. Votre formation intense et votre esprit de corps vous différencient. Qu'il ait été capturé par Delta Green, cela n'était pas votre première rencontre avec le diable.\"],\"EOS+bp\":[\"En tant que membre d'une force comme les Rangers de l'Armée des États-Unis, vous avez volontairement choisi un chemin plus difficile que les autres soldats. Vous avez passé des années dans les entraînements les plus difficiles du monde et servez maintenant sur les missions les plus dangereuses.\"],\"vS6YOc\":[\"Athlétique\"],\"JGpQgX\":[\"Athlétisme\"],\"BO9EJL\":[\"Éviter d'être touché\"],\"GDkUlj\":[\"Maladroite\"],\"ZuyHGl\":[\"Statistiques de base\"],\"o5DA10\":[\"Croyances\"],\"Bj0YCx\":[\"Plat\"],\"pFWzUl\":[\"Description du lien\"],\"joDnA9\":[\"Détails du lien\"],\"Qp5ZKm\":[\"Nom de lien\"],\"004WwA\":[\"Type de lien\"],\"MTMViL\":[\"Liens\"],\"ooT9hB\":[\"Détails des liens\"],\"YfLs0q\":[\"Les liens représentent des relations significatives que votre agent a avec les personnages non joueurs.\"],\"7zP6Gy\":[\"Les liens représentent les relations de votre agent avec les autres personnages.\\n                    Chaque lien commence avec un score égal à la Charisme de votre agent (\",[\"0\"],\").\\n                    Le score d'un lien peut s'améliorer si elle est nourrie entre les opérations. Un lien ne peut jamais avoir un score supérieur à la Charisme de votre agent.\\n                    Chaque fois que le Charisme de votre agent diminue, le score de chaque lien diminue de la même quantité.\\n                    Les liens peuvent être définis comme une personne ou un groupe de personnes. Par exemple, un agent peut avoir des liens avec son époux, son partenaire commercial et son ligue de bowling.\\n                    Un lien doit être une personne ou un petit groupe de personnes qui sont vivantes et peuvent être interagies avec.\"],\"PqoQ2W\":[\"Bonus\"],\"Mxgo1h\":[\"Points bonus\"],\"jw0SVP\":[\"Points bonus restants\"],\"MIHH6C\":[\"Point de rupture\"],\"jEvJsi\":[\"Brillant\"],\"GtWnxo\":[\"Brillant\"],\"Nr0SIL\":[\"Bureaucratie\"],\"CBGYlg\":[\"Mathématiques commerciales\"],\"6rScUj\":[\"Camping, tracking, improvisation dans la nature\"],\"a46V7b\":[\"Captivant\"],\"aw5Ht8\":[\"Captivité ou emprisonnement\"],\"5NsOuM\":[\"Chat-like\"],\"hVHY2j\":[\"Charisme\"],\"gLHd7f\":[\"Cocher la case d'un lien lorsque vous projettez des dégâts de SAN\"],\"DK/qpg\":[\"Cocher une compétence lorsque vous ratez un test de compétence. Après une session, ajoutez 1d4 à chaque compétence cochée et effacez le test.\"],\"nUIcQr\":[\"Choisir \",[\"0\"],\" compétences supplémentaires\"],\"zikS7b\":[\"Choisir un package de compétences bonus\"],\"k/xMy9\":[\"Choisir un package pour appliquer automatiquement les points bonus, ou sélectionner 'Aucun package' pour assigner manuellement vos points bonus\"],\"CGI4Ui\":[\"Église ou groupe de soutien\"],\"xCJdfg\":[\"Effacer\"],\"Jz0/oR\":[\"Effacer les descriptions\"],\"LndrJY\":[\"Effacer les compétences sélectionnées\"],\"Ql//nl\":[\"Effacer les modèles\"],\"OxOc2C\":[\"Maladroit\"],\"RLWPXk\":[\"Collègues dans un travail intense\"],\"KLEQcs\":[\"Diplôme principal ou années d'expérience; avec 50% en langue étrangère, vous avez la fluence native.\"],\"rIdFzg\":[\"Diplôme mineur ou entraînement de base.\"],\"yKoSLY\":[\"Captivant\"],\"kHbyNa\":[\"Informatique\"],\"AU0Kdy\":[\"Les ordinateurs et les machines sont l'épine dorsale de l'industrie moderne. Vous êtes un artisan des données ou des machines, possiblement pour le gouvernement et très certainement à but lucratif. Quelle que soit la façon dont vous utilisez vos compétences, le chevauchement entre les technologies de l'information et la conscience du surnaturel pourrait faire de ce métier le plus dangereux de la planète.\"],\"fOaHdd\":[\"Constitution\"],\"uhhJuh\":[\"Explosions contrôlées\"],\"IbtVsv\":[\"Convaincre les gens\"],\"3zF2gs\":[\"Couch potato\"],\"gtWP1l\":[\"Peut prendre les escaliers\"],\"A8wadO\":[\"Collègue ou partenaire\"],\"UTfQwZ\":[\"Artisanat\"],\"470Fk4\":[\"Artisanat (Macramé)\"],\"XjzLm3\":[\"Passions créatives\"],\"PMunmF\":[\"Criminel\"],\"oWJvZ7\":[\"Criminologie\"],\"VoIJ7Q\":[\"Investigation criminelle\"],\"5ALalD\":[\"Crédulité\"],\"d/rSvQ\":[\"Courage, logique, et intuition\"],\"Hp1l6f\":[\"Actuel\"],\"zIo0DE\":[\"Professions personnalisées\"],\"7/D6Vk\":[\"Quelqu'un qui essaie de tout faire.\"],\"iX6K3g\":[\"Modèles de vétéran blessé\"],\"7uHWR+\":[\"Date de naissance\"],\"onAw8T\":[\"Des décennies d'expérience, ou un diplôme de master ou de doctorat.\"],\"tffdBt\":[\"Dossier déclassifié\"],\"L3rQ+j\":[\"Agile\"],\"VVywaC\":[\"Délicat\"],\"PgSdTp\":[\"Démolition\"],\"ezghSE\":[\"Attributs dérivés\"],\"Nu4oKW\":[\"Description\"],\"s3Uzfx\":[\"Déterminé\"],\"08z1KE\":[\"Dextérité\"],\"u9l1aF\":[\"Dés\"],\"DwvMe2\":[\"Fainéant\"],\"BJ6YQv\":[\"Disguise\"],\"uPWSEh\":[\"Caractéristiques distinctives\"],\"9/Nzrk\":[\"Voulez-vous fermer les yeux à ce que vous avez vu et retourner dormir ? Ou venez avec ce burnout psychique et accomplissez l'impossible contre l'incroyable et maintenir le futur en retrait pour un autre jour ?\"],\"difbU5\":[\"Voulez-vous des drogues ?\"],\"DM17Gg\":[\"Les médecins sont souvent les premiers à découvrir les signes d'une incursion non naturelle, et les enquêteurs les plus précieux de ses effets dévastateurs sur l'humanité.\"],\"p+tvhp\":[\"Dévier\"],\"OUmX26\":[\"Dossier\"],\"Q2ObDC\":[\"Conduire\"],\"1afgyS\":[\"Dénué de toute énergie\"],\"aqxYLv\":[\"Éducation\"],\"n9u/m0\":[\"Employeur\"],\"pbUVnA\":[\"Captivant\"],\"iy5fSS\":[\"Ingénieur\"],\"RI/fzG\":[\"Entrez un sous-type\"],\"HmI5oK\":[\"Exemple\"],\"O4wxl8\":[\"Exporter en PDF\"],\"S8GwkS\":[\"Violence extrême\"],\"F7dRcr\":[\"Juste\"],\"+UN+Ez\":[\"Fatigué\"],\"y+uhgI\":[\"Agent fédéral\"],\"hrDJ2/\":[\"Faible\"],\"oCrptp\":[\"Trouver le chemin\"],\"qUl/bH\":[\"Trouver les choses\"],\"GnG+vN\":[\"Armes de feu\"],\"/OsYEc\":[\"Pompiers\"],\"RtEHo7\":[\"Ferme\"],\"11+tdk\":[\"Premiers soins\"],\"kODvZJ\":[\"Prénom\"],\"ZX5IeB\":[\"Première soirée à l'opéra ?\"],\"u4iXQi\":[\"Athlétique\"],\"wMLcWn\":[\"Flexible dans le principe\"],\"yaw+7x\":[\"Paramètres de police\"],\"kXTwNy\":[\"Langues étrangères\"],\"aFFQEb\":[\"Langues étrangères (Français)\"],\"a9k1oC\":[\"Officier de service étranger\"],\"z7q/DL\":[\"Forensique\"],\"K9cGgQ\":[\"Oubliable\"],\"0BlDs+\":[\"Forklifts, grues, excavateurs, etc\"],\"bdeAb2\":[\"Formidable\"],\"noN438\":[\"Fortifié\"],\"1Yrw83\":[\"Trouvé sur la page \",[\"pageNumber\"]],\"R2sFu3\":[\"Fragile\"],\"ZQfuYY\":[\"Faible\"],\"nLC6tu\":[\"Français\"],\"yTcD6q\":[\"Maladroite\"],\"ZJ1pST\":[\"Gene Parmesan, il est le meilleur\"],\"7o4S1E\":[\"Génie\"],\"+/PqzT\":[\"Maladroit\"],\"47ofD/\":[\"Éclair de glisse\"],\"fcFvf1\":[\"Lubrifier les roues de l'État\"],\"L8fEEm\":[\"Groupe\"],\"J36o9O\":[\"Pratique\"],\"fjd2D3\":[\"Expérience difficile\"],\"bzM4ak\":[\"Santé et robustesse\"],\"yGS9cI\":[\"Sain\"],\"UjSOv0\":[\"Robuste\"],\"NxZ7K+\":[\"Machines lourdes\"],\"Nw0C1v\":[\"Armes lourdes\"],\"0ZBnED\":[\"Inutilité\"],\"ORe4Lw\":[\"Historien\"],\"0caMy7\":[\"Histoire\"],\"Bwb/M2\":[\"Points de vie\"],\"1+WSY1\":[\"Loisirs\"],\"i0qMbr\":[\"Accueil\"],\"UU7+rG\":[\"Chevaux et autres\"],\"21qVcE\":[\"Gigantesque\"],\"Po1nwb\":[\"Renseignement humain\"],\"4bNpE0\":[\"Si la communauté du renseignement était une famille, Delta Green serait l'oncle dont personne ne parle.<0/>Delta Green est un Jeu de Rôle sur Table qui utilise un système D100. Les Gestionnaires dirigent des opérations pour leurs agents, d'autres joueurs, où une opération gouvernementale secrète œuvre pour protéger le monde des horreurs cosmiques inconnues.\"],\"dbLUe7\":[\"Si vous ne l'avez pas étudié, vous êtes condamné à le répéter\"],\"p64A7d\":[\"Si vous souhaitez soutenir ce travail, vous pouvez acheter du contenu Delta Green via des liens d'affiliation.\"],\"NNy6W8\":[\"Impulsif\"],\"b23dA7\":[\"Dans le FBI, le NSA et le CIA, il y a ceux qui collectent des informations et ceux qui décident de leur signification. Vous prenez des informations de sources diverses—journaux, sites web, informateurs, ELINT, et les actifs développés par les officiers de cas—et découvrez ce qu'ils signifient. En bref, votre travail est de mettre ensemble des connaissances sans lien, un entreprise dangereuse dans le monde de Delta Green.\"],\"UFi3x2\":[\"Incurieux\"],\"fByw/g\":[\"Individuel\"],\"+RRbXc\":[\"Indomptable\"],\"aemL3D\":[\"Affaibli\"],\"C5kxqk\":[\"Créer des miracles avec du ruban adhésif\"],\"s2/OPX\":[\"Insipide\"],\"Gzw2pq\":[\"Intelligence\"],\"Mf5ErF\":[\"Analyste d'intelligence\"],\"3FDXvB\":[\"Officier de cas d'intelligence\"],\"VGX26E\":[\"Irrationnel\"],\"Jp7JUr\":[\"C'est une chose de camion, vous ne le comprenez pas\"],\"mEzWf0\":[\"Jacked\"],\"tcBJo7\":[\"Sauter, esquiver, courir, grimper, etc\"],\"yoMIHR\":[\"Jurassic Park\"],\"sNoMVY\":[\"Statistiques clés\"],\"zC0r/K\":[\"Couteaux, haches, épées, etc\"],\"87a/t/\":[\"Étiquette\"],\"ok/7qO\":[\"Terre\"],\"UXBCwc\":[\"Nom de famille\"],\"iqhOTb\":[\"Loi\"],\"whBNst\":[\"Avocat\"],\"XHdFzA\":[\"Listez les motivations de votre personnage et ses troubles mentaux...\"],\"AqDRRD\":[\"LMGs que je pense\"],\"1z0Tzf\":[\"Macramé\"],\"kT7U6s\":[\"Magnétique\"],\"oLEThm\":[\"Faire exploser des choses à distance\"],\"PONy2U\":[\"Saisie manuelle\"],\"LlSUa5\":[\"De nombreux agents de Delta Green sont des officiers de la police fédérale, principalement du FBI. Delta Green a longtemps décidé que les agents fédéraux avaient le meilleur équilibre de compétences et de stabilité mentale nécessaire pour affronter le non naturel.\"],\"PAXVdT\":[\"Marine\"],\"CK1KXz\":[\"Maximum\"],\"spK3cv\":[\"Spécialiste des médias\"],\"zeVoXl\":[\"Pratique médicale\"],\"hS5cj7\":[\"Les professionnels de la médecine sont au front lorsque de terribles choses arrivent. C'est ce qui vous a amené à l'attention du groupe ?\"],\"6Kw9Gv\":[\"Médecine\"],\"F5rArt\":[\"Médiocre\"],\"syonIH\":[\"Retrouvez-moi au <0>Creuset</0> et parlez-moi un peu de vous, agent. Quelle est votre histoire ?\"],\"qx+7SQ\":[\"Armes de combat\"],\"e3Oj4X\":[\"Mesmérisant\"],\"2NOxz2\":[\"Initiale du milieu\"],\"dImRx9\":[\"Science militaire\"],\"YOtRZ0\":[\"Science militaire (Terre)\"],\"hNi0vh\":[\"Tactiques militaires\"],\"lJ71dZ\":[\"Soins médicaux urgents mineurs\"],\"dOLpsA\":[\"Motivations\"],\"aijpeY\":[\"Se déplacer dans l'eau\"],\"e9CT52\":[\"Mélanger les eaux d'un crime\"],\"av28MK\":[\"Musclé\"],\"2YKBHm\":[\"Nationalité\"],\"dbDjdV\":[\"Naviguer\"],\"mqUHTq\":[\"Presque immortel\"],\"Ysyt+a\":[\"Agile\"],\"kSzR8n\":[\"Aucun package (Points bonus manuels)\"],\"u3hw5g\":[\"Notez les détails supplémentaires sur le lien\"],\"XOGTWH\":[\"Remarquer les choses\"],\"9OGltc\":[\"Nombre de liens sociaux disponibles pour votre personnage.\"],\"tEPJOu\":[\"Infirmière\"],\"xhyJlX\":[\"Obsessions\"],\"7JOZLS\":[\"Occultisme\"],\"i0qI4Z\":[\"Conduire une voiture\"],\"ZI3xDG\":[\"Conduire un véhicule aérien\"],\"0zpgxV\":[\"Options\"],\"DW8nOh\":[\"Oraculaire\"],\"08AP9H\":[\"Peinture\"],\"pYwzJ9\":[\"Parlez-vous français ?\"],\"G/Cl3b\":[\"Paye pour Mensa\"],\"Xvtlo9\":[\"Persévérant\"],\"iLp4WI\":[\"Appel personnel et physique\"],\"DhAC6+\":[\"Données personnelles\"],\"sJYPeP\":[\"Détails personnels\"],\"8peLId\":[\"Motivation personnelle \",[\"0\"]],\"VpAfxU\":[\"Motivations personnelles\"],\"39qjNz\":[\"Les motivations personnelles représentent ce qui motive votre Agent au-delà de leurs liens. \\nCela peut être la foi, le patriotisme, les loisirs, ou même l'amour d'un animal de compagnie. \\nBien que puissantes, ces motivations ne sont pas aussi fortes que les liens avec d'autres humains.\"],\"9XnWrG\":[\"spécialité personnelle (n'importe quelle)\"],\"IzBQrW\":[\"Personnalité\"],\"MdYRcY\":[\"Convaincre\"],\"XdLRtZ\":[\"tir\"],\"BggFo8\":[\"Pharmacie\"],\"aUuLB9\":[\"Physique, chimie, biologie, etc\"],\"JRwb4O\":[\"Médecin\"],\"4+u/ep\":[\"Physique, chimie, biologie, etc\"],\"BVcN3k\":[\"Pilote\"],\"+R3h1p\":[\"Pilote (Aéronef)\"],\"YCHZ9w\":[\"Accueillant\"],\"idLVai\":[\"Flexible\"],\"8TEhqN\":[\"Achat par points\"],\"COUcFd\":[\"Points\"],\"BK4VZy\":[\"Officier de police\"],\"JoUIEx\":[\"Polymathie\"],\"NHfDC5\":[\"Pouvoir\"],\"n+1nKb\":[\"Compétences professionnelles prédéfinies\"],\"F6Ibbh\":[\"Profession & Compétences\"],\"pNZe3j\":[\"Détails de la profession\"],\"EiRYvW\":[\"Profession\"],\"fFTzL+\":[\"Directeur de programme\"],\"yGzhm3\":[\"Données psychologiques\"],\"45eqP/\":[\"Psychanalyse\"],\"VfxeQo\":[\"Coup, pied, grapple, morsure, etc\"],\"Xdl7I6\":[\"Radiant\"],\"NrA1Tk\":[\"Puissance physique brute\"],\"QKcpXG\":[\"Tranchant\"],\"37dxPM\":[\"Lit occasionnellement\"],\"YY/8OS\":[\"Statistiques recommandées pour votre profession choisie.\"],\"rYXfOA\":[\"Paramètres régionaux\"],\"XVLIk/\":[\"Enlever une balle, coudre une blessure\"],\"YUaj/8\":[\"Repoussant\"],\"/v4BCF\":[\"Réinitialiser les liens\"],\"nV3jkp\":[\"Réinitialiser les points bonus\"],\"oaVDgt\":[\"Réinitialiser tout\"],\"/x5uaa\":[\"Réinitialiser les détails personnels\"],\"MWfUCJ\":[\"Réinitialiser la profession\"],\"yKJGwK\":[\"Réinitialiser les points de compétence\"],\"e69RxD\":[\"Réinitialiser les statistiques\"],\"hBGyz8\":[\"Réinitialiser les statistiques\"],\"kogYC7\":[\"Résistant\"],\"rbfFz4\":[\"Déterminé\"],\"PcdWmG\":[\"Chevauchée\"],\"KZGIQJ\":[\"Robuste\"],\"maR6rE\":[\"Lancer 4d6, enlever le plus bas\"],\"dIRnNt\":[\"Sage\"],\"46u2kt\":[\"Marin\"],\"M+hL4r\":[\"Sanité\"],\"5/lolt\":[\"Science\"],\"4/rbeA\":[\"Science (Physique théorique)\"],\"RN1nHA\":[\"Scientifique\"],\"Eq6YVV\":[\"Note\"],\"A1taO8\":[\"Rechercher\"],\"PHOy1K\":[\"Sélectionner la date de naissance\"],\"oE8mTH\":[\"Sélectionner le thème\"],\"Tz0i8g\":[\"Paramètres\"],\"6L2EWT\":[\"Renseignement de signal\"],\"voEyf2\":[\"Renseignement de signal, cassage de codes\"],\"X1cBOB\":[\"Choix de compétences restants\"],\"ul6aFX\":[\"Points de compétence\"],\"e53RAl\":[\"Points de compétence restants\"],\"6gjFgQ\":[\"Note de compétence\"],\"professional-guidance.skill-cap-text\":[\"Les compétences peuvent atteindre \",[\"DEFAULT_MAX_SKILL_VALUE\"],\"% avec l'allocation de points, et atteindre \",[\"DEFAULT_TOTAL_CAP\"],\"% avec les points bonus. Si vos points bonus dépassaient \",[\"DEFAULT_TOTAL_CAP\"],\"%, la compétence atteindrait \",[\"DEFAULT_TOTAL_CAP\"],\"% et tous les points en excès seraient perdus. Les valeurs des compétences peuvent augmenter au-delà de \",[\"DEFAULT_TOTAL_CAP\"],\"% avec l'expérience acquise pendant le jeu.\"],\"aQnfn1\":[\"Se glisser\"],\"HXxT8R\":[\"Tant de choses sont illégales que des économies de crime existent. Ce profil correspond à un militant dur ou un criminel traditionnel “black collar” : pimp, cambrioleur, extorsionniste ou thug. Si vous voulez un criminel blanc-col, choisissez Scientifique des ordinateurs ou Directeur de l'entreprise et prenez des décisions très risquées.\"],\"bgjaGb\":[\"Mou\"],\"QLiQBs\":[\"Soldat\"],\"KNyHRH\":[\"Quelque chose que vous admirez\"],\"6jBuvW\":[\"Quelque chose que vous détestez\"],\"vsXP3+\":[\"Opérateur spécial\"],\"ITqzjt\":[\"Vitesse et agilité physique\"],\"9Cp1yM\":[\"Époux ou ex-époux\"],\"Pugge/\":[\"Professions standard\"],\"tXkhj/\":[\"Commencer\"],\"1ktYIh\":[\"commence à\"],\"24PEWv\":[\"Commence à:\"],\"DrzPEl\":[\"Données statistiques\"],\"2NbyY/\":[\"Statistiques\"],\"V2infW\":[\"Détails des statistiques\"],\"GU/TXk\":[\"Discrétion\"],\"KN4iFs\":[\"Rigide\"],\"JR0j+9\":[\"Force\"],\"6yvfxu\":[\"Stubborn\"],\"n3sBo6\":[\"Étude des humains et de leurs cultures\"],\"rKFfgg\":[\"Robuste\"],\"dXoieq\":[\"Résumé\"],\"XYLcNv\":[\"Support\"],\"nAmcJf\":[\"Chirurgie\"],\"QprZWI\":[\"Chirurgicalement précis\"],\"5ojEcZ\":[\"Survie\"],\"WWQ4IE\":[\"Susceptible\"],\"//qGGh\":[\"Nager\"],\"rk4kEq\":[\"Swole\"],\"niuT0f\":[\"Jetez un coup d'œil derrière le rideau. N'hésitez pas à :<0><1>signaler un <2>problème</2></1><3>laisser des <4>commentaires/demandes de fonctionnalités</4></3><5>ou ⭐️ Étoiler le projet pour booster l'ego de l'auteur.</5></0>\"],\"47lL4r\":[\"Persévérant\"],\"S4gbUS\":[\"Le Creuset\"],\"/X90px\":[\"Le nom de l'individu ou du groupe avec lequel votre agent est lié.\"],\"T4VBOv\":[\"Le titre de votre métier de l'agent\"],\"7bPwkx\":[\"Physique théorique\"],\"udFMBl\":[\"Épais\"],\"ndeZXv\":[\"Les choses que l'homme n'était pas destiné à savoir\"],\"P2YRxu\":[\"Titre :\"],\"aw0XRr\":[\"pour acheter les documents sources de Delta Green et en savoir plus.\"],\"72c5Qo\":[\"Total\"],\"u0Vxnx\":[\"Combat sans arme\"],\"6+uTR4\":[\"Inhabituel\"],\"tNJJOF\":[\"Maladroite\"],\"K+/GwL\":[\"Comprendre et argumenter la loi\"],\"l2ynLe\":[\"Comprendre le comportement humain\"],\"UyWBj5\":[\"Insain\"],\"Hs/oFo\":[\"Dénué de toute énergie\"],\"haSd2S\":[\"Irrationnel\"],\"xa8Jrn\":[\"Oubliable\"],\"BoY0Al\":[\"Indomptable\"],\"tI35bN\":[\"Utiliser toutes les suggestions\"],\"yjOG8K\":[\"Utiliser les suggestions\"],\"COU52/\":[\"Utilisez ces prompts pour vous imprégner de votre personnage. Elles sont toutes facultatives mais peuvent vous aider à vous mettre dans les pieds de votre personnage.\"],\"dsWcdt\":[\"Vigorant\"],\"K2f7tT\":[\"Violence\"],\"ItnLvW\":[\"Visitez\"],\"yudl3a\":[\"Nous menons une guerre contre l'inévitable. Et c'est une guerre que nous allons perdre. Mais nous gagnerons les batailles. Nous ferons travailler ces salauds pour ça. Nous les forcerons à prendre le terrain centimètre par centimètre. Et nous les ferons saigner pour chaque centimètre qu'ils prendront.\"],\"9ls/Vs\":[\"Faible\"],\"k6ZCLg\":[\"Faible-volontaire\"],\"Vo1ACK\":[\"Guerrier de fin de semaine\"],\"/oYuXu\":[\"Ce qui se passe à l'opéra reste ici. De toute façon, personne en dehors de l'organisation ne vous croirait.\"],\"P8Xrdt\":[\"Qu'est-ce que Delta Green ?\"],\"imD/Q0\":[\"Que vous soyez à bord d'un navire de guerre ou d'un navire marchand, vous avez passé votre vie sur les mers. Les mers cachent de nombreux secrets, et ceux qui y passent suffisamment de temps apprennent que certaines choses sont mieux laissées en paix. Quelque chose que vous avez rencontré dans ces vastes eaux a attiré l'attention de Delta Green — peut-être une chose qui ne devrait pas flotter, ou une découverte dans une profondeur où rien ne devrait survivre.\"],\"DUpGIu\":[\"Fougueux\"],\"RBd2c3\":[\"Pourquoi Delta Green fait-il confiance à cet agent ?\"],\"CyfgYa\":[\"Pourquoi cet agent sert-il Delta Green ?\"],\"imbYbz\":[\"Volonté\"],\"fDORhQ\":[\"Volonté, esprit et stabilité mentale\"],\"kP3ycB\":[\"Volonté\"],\"mYJzKG\":[\"x5\"],\"375wgc\":[\"Vous étendez les connaissances humaines dans un domaine comme la biologie, la physique ou la chimie. Lorsque certaines formes de connaissances provoquent de l'insanité et la mort, il est facile de conclure que certaines hypothèses ne devraient pas être testées.\"],\"vTWySo\":[\"Vous êtes un auteur, un éditeur, un chercheur pour une entreprise ou toutes branches du gouvernement, un bloggeur, un reporter de télévision ou un spécialiste des textes rares. Avec le non naturel, vous avez découvert l'histoire de toute une vie.\"],\"y87741\":[\"Vous reconstituez la vérité à partir de fragments du passé — documents, artefacts et histoires murmurées qui n'ont pas été intégrées aux manuels scolaires. Votre recherche vous a mené sur des chemins que l'académie mainstream refuse d'admettre, dévoilant des motifs dans l'histoire humaine qui suggèrent qu'il existe quelque chose de vaste et de terrible derrière la montée et la chute de la civilisation. Certains archives sont clos pour de bonnes raisons.\"],\"eh3cxY\":[\"Vous dirigez une organisation. Quelqu'un doit sécuriser les fonds, déplacer les ressources et faire des connexions, et c'est vous. Vous contrôlez un budget et êtes responsable de la maintenance de votre programme et du placement de l'argent. Les organisations découvrent les choses les plus surprenantes dans leur quête de profit ou du bien public.\"],\"V1GS2V\":[\"Vous servez et protégez. Les policiers de quartiers patrouillent en uniforme. Les shérifs adjoints répondent à un enforcé élu et ont compétence sur un comté entier. Les détectives arrivent après coup et reconstituent les éléments.\"],\"eYVzTc\":[\"Vous étudiez l'humanité dans toutes ses formes — de tribus isolées à sous-cultures urbaines. Votre travail vous mène dans les coins les plus reculés du monde où des traditions anciennes persistent et où le savoir interdit perdure. Que vous meniez des entrevues en régions en guerre ou que vous analysiez des artefacts culturels qui défient l'explication, vous avez appris que certaines traditions existent pour des raisons plus sombres que personne ne soupçonne.\"],\"doadht\":[\"Vous voyagez dans des terres étranges, rencontrez des personnes intéressantes et essayez de vous entendre avec elles. Les chances sont que vous travaillez pour le Département d'État, bien que USAID, le Service Commercial et le Service des Agriculture Étrangère aient également des FSOs. Quoi qu'il en soit, vous avez eu l'opportunité de connaître des choses exotiques et mortelles ; les choses qui vous qualifient pour la cléarance Delta Green.\"],\"65+vXE\":[\"Vous êtes un membre de l'Armée des États-Unis, de la Marine ou de l'Air Force. Vous avez été formé à la lutte et à la survie. Vous avez vu des choses qui font que le citoyen moyen crie.\"],\"blWU4y\":[\"Votre Agent peut avoir jusqu'à cinq motivations personnelles. Ajoutez-les lors de la création du personnage ou au fur et à mesure que votre histoire se développe. Chaque fois que votre Agent subit un point de rupture, retirez une motivation pour représenter leur traumatisme en croissance.\"],\"W0MqKV\":[\"Votre Agent gagne 10% dans la compétence Non naturel et ajoute +20% à l'Occult. Réduisez la SAN de votre Agent de son POW. Votre Agent gagne un nouveau trouble causé par le Non naturel (voir page 72). Réinitialisez le point de rupture de votre Agent à sa nouvelle SAN moins POW.\"],\"1nTbeT\":[\"Votre travail oscille entre la routine de l'entretien de votre matériel, l'excitation lorsque l'alarme finit par venir, et le travail de l'enquête sur un lieu après que le brouillard ait été dissipé. Si vous êtes impliqué avec Delta Green, vous avez clairement trébucheté dans quelque chose de plus grave qu'un incendie de maison.\"],\"+naB54\":[\"Vos outils sont un ordinateur et un smartphone. Vous pourriez déplacer des millions de dollars, ou des bits de données, ou les deux. Ou vous pourriez être un procureur, un avocat de la défense ou un juge.\"],\"cFe/mJ\":[\"Charisme\"]}")as Messages;
+/*eslint-disable*/import type{Messages}from"@lingui/core";export const messages=JSON.parse("{\"stat.reminder\":[[\"reminder\"]],\"stat.label\":[[\"stat\"]],\"stat.intelligence.reminder\":[\"Cunning, logic, and intuition\"],\"stat.constitution.reminder\":[\"Health and hardiness\"],\"stat.charisma.reminder\":[\"Personal and physical appeal\"],\"stat.strength.reminder\":[\"Raw physical power\"],\"stat.dexterity.reminder\":[\"Speed and physical agility\"],\"stat.power.reminder\":[\"Will, spirit, and mental stability\"],\"8VjWl0\":[\"--- Professions supplémentaires ---\"],\"KE5vnC\":[\"--- Professions standard ---\"],\"7o9kk2\":[\"(Choisir un type ci-dessous)\"],\"J/hVSQ\":[[\"0\"]],\"X3V8cR\":[[\"0\"]],\"xYxQCZ\":[[\"0\"],\" \",[\"1\"]],\"Zz4Qvw\":[[\"capitalizedStat\"]],\"w9dLMW\":[[\"label\"]],\"JpzPnP\":[[\"labelText\"]],\"choose-skills.no-choices\":[[\"professionName\"],\" n'a pas de compétences flexibles à choisir.\"],\"leNkYU\":[[\"reminderText\"]],\"5ft2yB\":[[\"skillCapText\"]],\"kDmBhk\":[[\"statKey\"]],\"urTZDx\":[[\"statKey\"]],\"LZJAen\":[\"`$\",[\"0\"],\" has no flexible skills to choose from.`\"],\"9eGA8Z\":[\"010111100001\"],\"Hzaa8H\":[\"1. Nom, Prénom, Initiale du milieu\"],\"yQJwH3\":[\"10. Description physique\"],\"kmhPgQ\":[\"11. Liens\"],\"yGWS1j\":[\"12. Motivations et troubles mentaux\"],\"w/Xy4v\":[\"13. Incidents de perte de SAN sans devenir fou\"],\"VnPSmP\":[\"2. Profession (Grade si applicable)\"],\"qVQ1QT\":[\"3. Employeur\"],\"nVobO+\":[\"4. Nationalité\"],\"icgCRe\":[\"5. Sexe\"],\"+18IK1\":[\"6. Âge et date de naissance\"],\"AbZcNQ\":[\"7. Formation et histoire professionnelle\"],\"bzqTDJ\":[\"8. Statistiques\"],\"W5FXNn\":[\"9. Attributs dérivés\"],\"5XYRnE\":[\"Un amateur dévoué; avec une langue étrangère, vous pouvez avoir des conversations rudimentaires.\"],\"igi2jT\":[\"Une poursuite de toute une vie ou plusieurs doctorats.\"],\"uyJsf6\":[\"À propos\"],\"2zPUa5\":[\"À propos de l'auteur\"],\"UNDRet\":[\"À propos du jeu\"],\"WcvywZ\":[\"Comptabilité\"],\"F6pfE9\":[\"Actif\"],\"zq0AxB\":[\"Actif couch potato\"],\"yGwMqg\":[\"adapté\"],\"0kRNUo\":[\"Ajoutez +10% à la compétence Non naturel de votre Agent et +10% à cinq compétences autres que Non naturel. Cela ne peut pas amener une compétence à plus de 90%. Réduisez la SAN de votre Agent de 5. Retirez un lien.\"],\"Xi6Is8\":[\"Ajoutez +10% à la compétence Non naturel de votre Agent. Réduisez la SAN de 5. Retirez 3 de la CHA de votre Agent et de chaque lien. Votre Agent est adapté à la violence (voir page 73).\"],\"EHRBpm\":[\"Ajoutez +10% à la compétence Non naturel de votre Agent. Réduisez la SAN de 5. Retirez 3 de la POW de votre Agent. Votre Agent est adapté à l'impuissance (voir page 73).\"],\"custom-skill-form.bonus-points-reminder\":[\"Ajoute \",[\"DEFAULT_BONUS_VALUE\"],\" à la compétence totale pour chaque point bonus alloué (limité à \",[\"DEFAULT_TOTAL_CAP\"],\")\"],\"7TlSPN\":[\"Adroit\"],\"zVgowR\":[\"Affable\"],\"CmUtio\":[\"Agile\"],\"gXy38H\":[\"Air ou mer, commercial ou militaire, votre devoir est de maintenir vos passagers en vie et votre bateau intact. Cela peut mener à des choix difficiles lorsque vos passagers mettent le véhicule en danger. Ou êtes-vous un opérateur de drone, pilotant un Predator à mille miles de distance ? Quoi qu'il en soit, quelle opération vous a fait remarquer à Delta Green ?\"],\"Z+c4we\":[\"Aéronef\"],\"rPKlAj\":[\"Alerte\"],\"EbuTi3\":[\"Pseudonyme\"],\"JIeg5e\":[\"Tous les doigts\"],\"6zd6CL\":[\"Attirant\"],\"UCIiuP\":[\"Analyse du raisonnement\"],\"WbrlFU\":[\"Andrew Sobiesiak est un ingénieur logiciel qui joue aux JDR sur table en ligne chaque semaine depuis 2017.<0/> Il ferait un terrible agent de terrain mais est prêt à combattre la guerre cosmique sisyphéenne. Il vit à <1>Seattle, WA</1> avec sa <2>femme, 5 chats et un chien</2>\"],\"2qgwCU\":[\"Anthropologue\"],\"juHsSO\":[\"Anthropologie\"],\"aAIQg2\":[\"Apparence\"],\"f+WR1d\":[\"Compétences applicables\"],\"FGdRNI\":[\"Appliquer les compétences prédéfinies et choisir des compétences supplémentaires pour votre personnage.\"],\"xLXsjU\":[\"Arc Dream Publishing est un petit éditeur qui produit des jeux de haute qualité, notamment des jeux de rôle, un magazine de jeux d'horreur et des jeux de cartes. Parmi leurs œuvres populaires figurent Delta Green, Swords & Sorceries et The Black Company RPG.\"],\"vJo5g6\":[\"Archéologie\"],\"+QARA4\":[\"Art\"],\"Zvntde\":[\"Art (Peinture)\"],\"ziY9j1\":[\"Artillerie\"],\"3Vj2Gv\":[\"En tant que membre des Marines des États-Unis, vous faites partie de la force expéditionnaire élite des États-Unis en état de préparation. Premier à combattre, vous avez été formé à opérer dans n'importe quelle environnement, de la guerre urbaine à des opérations amphibies. Votre formation intense et votre esprit de corps vous différencient. Qu'il ait été capturé par Delta Green, cela n'était pas votre première rencontre avec le diable.\"],\"EOS+bp\":[\"En tant que membre d'une force comme les Rangers de l'Armée des États-Unis, vous avez volontairement choisi un chemin plus difficile que les autres soldats. Vous avez passé des années dans les entraînements les plus difficiles du monde et servez maintenant sur les missions les plus dangereuses.\"],\"vS6YOc\":[\"Athlétique\"],\"JGpQgX\":[\"Athlétisme\"],\"BO9EJL\":[\"Éviter d'être touché\"],\"GDkUlj\":[\"Maladroite\"],\"ZuyHGl\":[\"Statistiques de base\"],\"o5DA10\":[\"Croyances\"],\"Bj0YCx\":[\"Plat\"],\"pFWzUl\":[\"Description du lien\"],\"joDnA9\":[\"Détails du lien\"],\"Qp5ZKm\":[\"Nom de lien\"],\"004WwA\":[\"Type de lien\"],\"MTMViL\":[\"Liens\"],\"ooT9hB\":[\"Détails des liens\"],\"YfLs0q\":[\"Les liens représentent des relations significatives que votre agent a avec les personnages non joueurs.\"],\"7zP6Gy\":[\"Les liens représentent les relations de votre agent avec les autres personnages.\\n                    Chaque lien commence avec un score égal à la Charisme de votre agent (\",[\"0\"],\").\\n                    Le score d'un lien peut s'améliorer si elle est nourrie entre les opérations. Un lien ne peut jamais avoir un score supérieur à la Charisme de votre agent.\\n                    Chaque fois que le Charisme de votre agent diminue, le score de chaque lien diminue de la même quantité.\\n                    Les liens peuvent être définis comme une personne ou un groupe de personnes. Par exemple, un agent peut avoir des liens avec son époux, son partenaire commercial et son ligue de bowling.\\n                    Un lien doit être une personne ou un petit groupe de personnes qui sont vivantes et peuvent être interagies avec.\"],\"PqoQ2W\":[\"Bonus\"],\"Mxgo1h\":[\"Points bonus\"],\"jw0SVP\":[\"Points bonus restants\"],\"MIHH6C\":[\"Point de rupture\"],\"jEvJsi\":[\"Brillant\"],\"GtWnxo\":[\"Brillant\"],\"Nr0SIL\":[\"Bureaucratie\"],\"CBGYlg\":[\"Mathématiques commerciales\"],\"6rScUj\":[\"Camping, tracking, improvisation dans la nature\"],\"a46V7b\":[\"Captivant\"],\"aw5Ht8\":[\"Captivité ou emprisonnement\"],\"5NsOuM\":[\"Chat-like\"],\"hVHY2j\":[\"Charisme\"],\"gLHd7f\":[\"Cocher la case d'un lien lorsque vous projettez des dégâts de SAN\"],\"DK/qpg\":[\"Cocher une compétence lorsque vous ratez un test de compétence. Après une session, ajoutez 1d4 à chaque compétence cochée et effacez le test.\"],\"nUIcQr\":[\"Choisir \",[\"0\"],\" compétences supplémentaires\"],\"zikS7b\":[\"Choisir un package de compétences bonus\"],\"k/xMy9\":[\"Choisir un package pour appliquer automatiquement les points bonus, ou sélectionner 'Aucun package' pour assigner manuellement vos points bonus\"],\"CGI4Ui\":[\"Église ou groupe de soutien\"],\"xCJdfg\":[\"Effacer\"],\"Jz0/oR\":[\"Effacer les descriptions\"],\"LndrJY\":[\"Effacer les compétences sélectionnées\"],\"Ql//nl\":[\"Effacer les modèles\"],\"OxOc2C\":[\"Maladroit\"],\"RLWPXk\":[\"Collègues dans un travail intense\"],\"KLEQcs\":[\"Diplôme principal ou années d'expérience; avec 50% en langue étrangère, vous avez la fluence native.\"],\"rIdFzg\":[\"Diplôme mineur ou entraînement de base.\"],\"yKoSLY\":[\"Captivant\"],\"kHbyNa\":[\"Informatique\"],\"AU0Kdy\":[\"Les ordinateurs et les machines sont l'épine dorsale de l'industrie moderne. Vous êtes un artisan des données ou des machines, possiblement pour le gouvernement et très certainement à but lucratif. Quelle que soit la façon dont vous utilisez vos compétences, le chevauchement entre les technologies de l'information et la conscience du surnaturel pourrait faire de ce métier le plus dangereux de la planète.\"],\"fOaHdd\":[\"Constitution\"],\"uhhJuh\":[\"Explosions contrôlées\"],\"IbtVsv\":[\"Convaincre les gens\"],\"3zF2gs\":[\"Couch potato\"],\"gtWP1l\":[\"Peut prendre les escaliers\"],\"A8wadO\":[\"Collègue ou partenaire\"],\"UTfQwZ\":[\"Artisanat\"],\"470Fk4\":[\"Artisanat (Macramé)\"],\"XjzLm3\":[\"Passions créatives\"],\"PMunmF\":[\"Criminel\"],\"oWJvZ7\":[\"Criminologie\"],\"VoIJ7Q\":[\"Investigation criminelle\"],\"5ALalD\":[\"Crédulité\"],\"d/rSvQ\":[\"Courage, logique, et intuition\"],\"Hp1l6f\":[\"Actuel\"],\"zIo0DE\":[\"Professions personnalisées\"],\"7/D6Vk\":[\"Quelqu'un qui essaie de tout faire.\"],\"iX6K3g\":[\"Modèles de vétéran blessé\"],\"7uHWR+\":[\"Date de naissance\"],\"onAw8T\":[\"Des décennies d'expérience, ou un diplôme de master ou de doctorat.\"],\"tffdBt\":[\"Dossier déclassifié\"],\"L3rQ+j\":[\"Agile\"],\"VVywaC\":[\"Délicat\"],\"PgSdTp\":[\"Démolition\"],\"ezghSE\":[\"Attributs dérivés\"],\"Nu4oKW\":[\"Description\"],\"s3Uzfx\":[\"Déterminé\"],\"08z1KE\":[\"Dextérité\"],\"u9l1aF\":[\"Dés\"],\"DwvMe2\":[\"Fainéant\"],\"BJ6YQv\":[\"Disguise\"],\"uPWSEh\":[\"Caractéristiques distinctives\"],\"9/Nzrk\":[\"Voulez-vous fermer les yeux à ce que vous avez vu et retourner dormir ? Ou venez avec ce burnout psychique et accomplissez l'impossible contre l'incroyable et maintenir le futur en retrait pour un autre jour ?\"],\"difbU5\":[\"Voulez-vous des drogues ?\"],\"DM17Gg\":[\"Les médecins sont souvent les premiers à découvrir les signes d'une incursion non naturelle, et les enquêteurs les plus précieux de ses effets dévastateurs sur l'humanité.\"],\"p+tvhp\":[\"Dévier\"],\"OUmX26\":[\"Dossier\"],\"Q2ObDC\":[\"Conduire\"],\"1afgyS\":[\"Dénué de toute énergie\"],\"aqxYLv\":[\"Éducation\"],\"n9u/m0\":[\"Employeur\"],\"pbUVnA\":[\"Captivant\"],\"iy5fSS\":[\"Ingénieur\"],\"RI/fzG\":[\"Entrez un sous-type\"],\"HmI5oK\":[\"Exemple\"],\"O4wxl8\":[\"Exporter en PDF\"],\"S8GwkS\":[\"Violence extrême\"],\"F7dRcr\":[\"Juste\"],\"+UN+Ez\":[\"Fatigué\"],\"y+uhgI\":[\"Agent fédéral\"],\"hrDJ2/\":[\"Faible\"],\"oCrptp\":[\"Trouver le chemin\"],\"qUl/bH\":[\"Trouver les choses\"],\"GnG+vN\":[\"Armes de feu\"],\"/OsYEc\":[\"Pompiers\"],\"RtEHo7\":[\"Ferme\"],\"11+tdk\":[\"Premiers soins\"],\"kODvZJ\":[\"Prénom\"],\"ZX5IeB\":[\"Première soirée à l'opéra ?\"],\"u4iXQi\":[\"Athlétique\"],\"wMLcWn\":[\"Flexible dans le principe\"],\"yaw+7x\":[\"Paramètres de police\"],\"kXTwNy\":[\"Langues étrangères\"],\"aFFQEb\":[\"Langues étrangères (Français)\"],\"a9k1oC\":[\"Officier de service étranger\"],\"z7q/DL\":[\"Forensique\"],\"K9cGgQ\":[\"Oubliable\"],\"0BlDs+\":[\"Forklifts, grues, excavateurs, etc\"],\"bdeAb2\":[\"Formidable\"],\"noN438\":[\"Fortifié\"],\"1Yrw83\":[\"Trouvé sur la page \",[\"pageNumber\"]],\"R2sFu3\":[\"Fragile\"],\"ZQfuYY\":[\"Faible\"],\"nLC6tu\":[\"Français\"],\"yTcD6q\":[\"Maladroite\"],\"ZJ1pST\":[\"Gene Parmesan, il est le meilleur\"],\"7o4S1E\":[\"Génie\"],\"+/PqzT\":[\"Maladroit\"],\"47ofD/\":[\"Éclair de glisse\"],\"fcFvf1\":[\"Lubrifier les roues de l'État\"],\"L8fEEm\":[\"Groupe\"],\"J36o9O\":[\"Pratique\"],\"fjd2D3\":[\"Expérience difficile\"],\"bzM4ak\":[\"Santé et robustesse\"],\"yGS9cI\":[\"Sain\"],\"UjSOv0\":[\"Robuste\"],\"NxZ7K+\":[\"Machines lourdes\"],\"Nw0C1v\":[\"Armes lourdes\"],\"0ZBnED\":[\"Inutilité\"],\"ORe4Lw\":[\"Historien\"],\"0caMy7\":[\"Histoire\"],\"Bwb/M2\":[\"Points de vie\"],\"1+WSY1\":[\"Loisirs\"],\"i0qMbr\":[\"Accueil\"],\"UU7+rG\":[\"Chevaux et autres\"],\"21qVcE\":[\"Gigantesque\"],\"Po1nwb\":[\"Renseignement humain\"],\"4bNpE0\":[\"Si la communauté du renseignement était une famille, Delta Green serait l'oncle dont personne ne parle.<0/>Delta Green est un Jeu de Rôle sur Table qui utilise un système D100. Les Gestionnaires dirigent des opérations pour leurs agents, d'autres joueurs, où une opération gouvernementale secrète œuvre pour protéger le monde des horreurs cosmiques inconnues.\"],\"dbLUe7\":[\"Si vous ne l'avez pas étudié, vous êtes condamné à le répéter\"],\"p64A7d\":[\"Si vous souhaitez soutenir ce travail, vous pouvez acheter du contenu Delta Green via des liens d'affiliation.\"],\"NNy6W8\":[\"Impulsif\"],\"b23dA7\":[\"Dans le FBI, le NSA et le CIA, il y a ceux qui collectent des informations et ceux qui décident de leur signification. Vous prenez des informations de sources diverses—journaux, sites web, informateurs, ELINT, et les actifs développés par les officiers de cas—et découvrez ce qu'ils signifient. En bref, votre travail est de mettre ensemble des connaissances sans lien, un entreprise dangereuse dans le monde de Delta Green.\"],\"UFi3x2\":[\"Incurieux\"],\"fByw/g\":[\"Individuel\"],\"+RRbXc\":[\"Indomptable\"],\"aemL3D\":[\"Affaibli\"],\"C5kxqk\":[\"Créer des miracles avec du ruban adhésif\"],\"s2/OPX\":[\"Insipide\"],\"Gzw2pq\":[\"Intelligence\"],\"Mf5ErF\":[\"Analyste d'intelligence\"],\"3FDXvB\":[\"Officier de cas d'intelligence\"],\"VGX26E\":[\"Irrationnel\"],\"Jp7JUr\":[\"C'est une chose de camion, vous ne le comprenez pas\"],\"mEzWf0\":[\"Jacked\"],\"tcBJo7\":[\"Sauter, esquiver, courir, grimper, etc\"],\"yoMIHR\":[\"Jurassic Park\"],\"sNoMVY\":[\"Statistiques clés\"],\"zC0r/K\":[\"Couteaux, haches, épées, etc\"],\"87a/t/\":[\"Étiquette\"],\"ok/7qO\":[\"Terre\"],\"UXBCwc\":[\"Nom de famille\"],\"iqhOTb\":[\"Loi\"],\"whBNst\":[\"Avocat\"],\"XHdFzA\":[\"Listez les motivations de votre personnage et ses troubles mentaux...\"],\"AqDRRD\":[\"LMGs que je pense\"],\"1z0Tzf\":[\"Macramé\"],\"kT7U6s\":[\"Magnétique\"],\"oLEThm\":[\"Faire exploser des choses à distance\"],\"PONy2U\":[\"Saisie manuelle\"],\"LlSUa5\":[\"De nombreux agents de Delta Green sont des officiers de la police fédérale, principalement du FBI. Delta Green a longtemps décidé que les agents fédéraux avaient le meilleur équilibre de compétences et de stabilité mentale nécessaire pour affronter le non naturel.\"],\"PAXVdT\":[\"Marine\"],\"CK1KXz\":[\"Maximum\"],\"spK3cv\":[\"Spécialiste des médias\"],\"zeVoXl\":[\"Pratique médicale\"],\"hS5cj7\":[\"Les professionnels de la médecine sont au front lorsque de terribles choses arrivent. C'est ce qui vous a amené à l'attention du groupe ?\"],\"6Kw9Gv\":[\"Médecine\"],\"F5rArt\":[\"Médiocre\"],\"syonIH\":[\"Retrouvez-moi au <0>Creuset</0> et parlez-moi un peu de vous, agent. Quelle est votre histoire ?\"],\"qx+7SQ\":[\"Armes de combat\"],\"e3Oj4X\":[\"Mesmérisant\"],\"2NOxz2\":[\"Initiale du milieu\"],\"dImRx9\":[\"Science militaire\"],\"YOtRZ0\":[\"Science militaire (Terre)\"],\"hNi0vh\":[\"Tactiques militaires\"],\"lJ71dZ\":[\"Soins médicaux urgents mineurs\"],\"dOLpsA\":[\"Motivations\"],\"aijpeY\":[\"Se déplacer dans l'eau\"],\"e9CT52\":[\"Mélanger les eaux d'un crime\"],\"av28MK\":[\"Musclé\"],\"2YKBHm\":[\"Nationalité\"],\"dbDjdV\":[\"Naviguer\"],\"mqUHTq\":[\"Presque immortel\"],\"Ysyt+a\":[\"Agile\"],\"kSzR8n\":[\"Aucun package (Points bonus manuels)\"],\"u3hw5g\":[\"Notez les détails supplémentaires sur le lien\"],\"XOGTWH\":[\"Remarquer les choses\"],\"9OGltc\":[\"Nombre de liens sociaux disponibles pour votre personnage.\"],\"tEPJOu\":[\"Infirmière\"],\"xhyJlX\":[\"Obsessions\"],\"7JOZLS\":[\"Occultisme\"],\"i0qI4Z\":[\"Conduire une voiture\"],\"ZI3xDG\":[\"Conduire un véhicule aérien\"],\"0zpgxV\":[\"Options\"],\"DW8nOh\":[\"Oraculaire\"],\"08AP9H\":[\"Peinture\"],\"pYwzJ9\":[\"Parlez-vous français ?\"],\"G/Cl3b\":[\"Paye pour Mensa\"],\"Xvtlo9\":[\"Persévérant\"],\"iLp4WI\":[\"Appel personnel et physique\"],\"DhAC6+\":[\"Données personnelles\"],\"sJYPeP\":[\"Détails personnels\"],\"8peLId\":[\"Motivation personnelle \",[\"0\"]],\"VpAfxU\":[\"Motivations personnelles\"],\"6dVMln\":[\"Personal Motivations represent what drives your Agent beyond their Bonds. \\n    These can be faith, patriotism, hobbies, or even the love of a pet. While \\n    powerful, these motivations aren't as strong as Bonds to other humans.\"],\"39qjNz\":[\"Les motivations personnelles représentent ce qui motive votre Agent au-delà de leurs liens. \\nCela peut être la foi, le patriotisme, les loisirs, ou même l'amour d'un animal de compagnie. \\nBien que puissantes, ces motivations ne sont pas aussi fortes que les liens avec d'autres humains.\"],\"9XnWrG\":[\"spécialité personnelle (n'importe quelle)\"],\"IzBQrW\":[\"Personnalité\"],\"MdYRcY\":[\"Convaincre\"],\"XdLRtZ\":[\"tir\"],\"BggFo8\":[\"Pharmacie\"],\"aUuLB9\":[\"Physique, chimie, biologie, etc\"],\"JRwb4O\":[\"Médecin\"],\"4+u/ep\":[\"Physique, chimie, biologie, etc\"],\"BVcN3k\":[\"Pilote\"],\"+R3h1p\":[\"Pilote (Aéronef)\"],\"YCHZ9w\":[\"Accueillant\"],\"idLVai\":[\"Flexible\"],\"8TEhqN\":[\"Achat par points\"],\"COUcFd\":[\"Points\"],\"BK4VZy\":[\"Officier de police\"],\"JoUIEx\":[\"Polymathie\"],\"NHfDC5\":[\"Pouvoir\"],\"n+1nKb\":[\"Compétences professionnelles prédéfinies\"],\"F6Ibbh\":[\"Profession & Compétences\"],\"pNZe3j\":[\"Détails de la profession\"],\"EiRYvW\":[\"Profession\"],\"fFTzL+\":[\"Directeur de programme\"],\"yGzhm3\":[\"Données psychologiques\"],\"45eqP/\":[\"Psychanalyse\"],\"VfxeQo\":[\"Coup, pied, grapple, morsure, etc\"],\"Xdl7I6\":[\"Radiant\"],\"NrA1Tk\":[\"Puissance physique brute\"],\"QKcpXG\":[\"Tranchant\"],\"37dxPM\":[\"Lit occasionnellement\"],\"YY/8OS\":[\"Statistiques recommandées pour votre profession choisie.\"],\"rYXfOA\":[\"Paramètres régionaux\"],\"XVLIk/\":[\"Enlever une balle, coudre une blessure\"],\"YUaj/8\":[\"Repoussant\"],\"/v4BCF\":[\"Réinitialiser les liens\"],\"nV3jkp\":[\"Réinitialiser les points bonus\"],\"oaVDgt\":[\"Réinitialiser tout\"],\"/x5uaa\":[\"Réinitialiser les détails personnels\"],\"MWfUCJ\":[\"Réinitialiser la profession\"],\"yKJGwK\":[\"Réinitialiser les points de compétence\"],\"e69RxD\":[\"Réinitialiser les statistiques\"],\"hBGyz8\":[\"Réinitialiser les statistiques\"],\"kogYC7\":[\"Résistant\"],\"rbfFz4\":[\"Déterminé\"],\"PcdWmG\":[\"Chevauchée\"],\"KZGIQJ\":[\"Robuste\"],\"maR6rE\":[\"Lancer 4d6, enlever le plus bas\"],\"dIRnNt\":[\"Sage\"],\"46u2kt\":[\"Marin\"],\"M+hL4r\":[\"Sanité\"],\"5/lolt\":[\"Science\"],\"4/rbeA\":[\"Science (Physique théorique)\"],\"RN1nHA\":[\"Scientifique\"],\"Eq6YVV\":[\"Note\"],\"A1taO8\":[\"Rechercher\"],\"PHOy1K\":[\"Sélectionner la date de naissance\"],\"oE8mTH\":[\"Sélectionner le thème\"],\"Tz0i8g\":[\"Paramètres\"],\"6L2EWT\":[\"Renseignement de signal\"],\"voEyf2\":[\"Renseignement de signal, cassage de codes\"],\"X1cBOB\":[\"Choix de compétences restants\"],\"ul6aFX\":[\"Points de compétence\"],\"e53RAl\":[\"Points de compétence restants\"],\"6gjFgQ\":[\"Note de compétence\"],\"professional-guidance.skill-cap-text\":[\"Les compétences peuvent atteindre \",[\"DEFAULT_MAX_SKILL_VALUE\"],\"% avec l'allocation de points, et atteindre \",[\"DEFAULT_TOTAL_CAP\"],\"% avec les points bonus. Si vos points bonus dépassaient \",[\"DEFAULT_TOTAL_CAP\"],\"%, la compétence atteindrait \",[\"DEFAULT_TOTAL_CAP\"],\"% et tous les points en excès seraient perdus. Les valeurs des compétences peuvent augmenter au-delà de \",[\"DEFAULT_TOTAL_CAP\"],\"% avec l'expérience acquise pendant le jeu.\"],\"aQnfn1\":[\"Se glisser\"],\"HXxT8R\":[\"Tant de choses sont illégales que des économies de crime existent. Ce profil correspond à un militant dur ou un criminel traditionnel “black collar” : pimp, cambrioleur, extorsionniste ou thug. Si vous voulez un criminel blanc-col, choisissez Scientifique des ordinateurs ou Directeur de l'entreprise et prenez des décisions très risquées.\"],\"bgjaGb\":[\"Mou\"],\"QLiQBs\":[\"Soldat\"],\"KNyHRH\":[\"Quelque chose que vous admirez\"],\"6jBuvW\":[\"Quelque chose que vous détestez\"],\"vsXP3+\":[\"Opérateur spécial\"],\"ITqzjt\":[\"Vitesse et agilité physique\"],\"9Cp1yM\":[\"Époux ou ex-époux\"],\"Pugge/\":[\"Professions standard\"],\"tXkhj/\":[\"Commencer\"],\"1ktYIh\":[\"commence à\"],\"24PEWv\":[\"Commence à:\"],\"DrzPEl\":[\"Données statistiques\"],\"2NbyY/\":[\"Statistiques\"],\"V2infW\":[\"Détails des statistiques\"],\"GU/TXk\":[\"Discrétion\"],\"KN4iFs\":[\"Rigide\"],\"JR0j+9\":[\"Force\"],\"6yvfxu\":[\"Stubborn\"],\"n3sBo6\":[\"Étude des humains et de leurs cultures\"],\"rKFfgg\":[\"Robuste\"],\"dXoieq\":[\"Résumé\"],\"XYLcNv\":[\"Support\"],\"nAmcJf\":[\"Chirurgie\"],\"QprZWI\":[\"Chirurgicalement précis\"],\"5ojEcZ\":[\"Survie\"],\"WWQ4IE\":[\"Susceptible\"],\"//qGGh\":[\"Nager\"],\"rk4kEq\":[\"Swole\"],\"niuT0f\":[\"Jetez un coup d'œil derrière le rideau. N'hésitez pas à :<0><1>signaler un <2>problème</2></1><3>laisser des <4>commentaires/demandes de fonctionnalités</4></3><5>ou ⭐️ Étoiler le projet pour booster l'ego de l'auteur.</5></0>\"],\"47lL4r\":[\"Persévérant\"],\"S4gbUS\":[\"Le Creuset\"],\"/X90px\":[\"Le nom de l'individu ou du groupe avec lequel votre agent est lié.\"],\"T4VBOv\":[\"Le titre de votre métier de l'agent\"],\"7bPwkx\":[\"Physique théorique\"],\"udFMBl\":[\"Épais\"],\"ndeZXv\":[\"Les choses que l'homme n'était pas destiné à savoir\"],\"P2YRxu\":[\"Titre :\"],\"aw0XRr\":[\"pour acheter les documents sources de Delta Green et en savoir plus.\"],\"72c5Qo\":[\"Total\"],\"u0Vxnx\":[\"Combat sans arme\"],\"6+uTR4\":[\"Inhabituel\"],\"tNJJOF\":[\"Maladroite\"],\"K+/GwL\":[\"Comprendre et argumenter la loi\"],\"l2ynLe\":[\"Comprendre le comportement humain\"],\"UyWBj5\":[\"Insain\"],\"Hs/oFo\":[\"Dénué de toute énergie\"],\"haSd2S\":[\"Irrationnel\"],\"xa8Jrn\":[\"Oubliable\"],\"BoY0Al\":[\"Indomptable\"],\"tI35bN\":[\"Utiliser toutes les suggestions\"],\"yjOG8K\":[\"Utiliser les suggestions\"],\"COU52/\":[\"Utilisez ces prompts pour vous imprégner de votre personnage. Elles sont toutes facultatives mais peuvent vous aider à vous mettre dans les pieds de votre personnage.\"],\"dsWcdt\":[\"Vigorant\"],\"K2f7tT\":[\"Violence\"],\"ItnLvW\":[\"Visitez\"],\"yudl3a\":[\"Nous menons une guerre contre l'inévitable. Et c'est une guerre que nous allons perdre. Mais nous gagnerons les batailles. Nous ferons travailler ces salauds pour ça. Nous les forcerons à prendre le terrain centimètre par centimètre. Et nous les ferons saigner pour chaque centimètre qu'ils prendront.\"],\"9ls/Vs\":[\"Faible\"],\"k6ZCLg\":[\"Faible-volontaire\"],\"Vo1ACK\":[\"Guerrier de fin de semaine\"],\"/oYuXu\":[\"Ce qui se passe à l'opéra reste ici. De toute façon, personne en dehors de l'organisation ne vous croirait.\"],\"P8Xrdt\":[\"Qu'est-ce que Delta Green ?\"],\"imD/Q0\":[\"Que vous soyez à bord d'un navire de guerre ou d'un navire marchand, vous avez passé votre vie sur les mers. Les mers cachent de nombreux secrets, et ceux qui y passent suffisamment de temps apprennent que certaines choses sont mieux laissées en paix. Quelque chose que vous avez rencontré dans ces vastes eaux a attiré l'attention de Delta Green — peut-être une chose qui ne devrait pas flotter, ou une découverte dans une profondeur où rien ne devrait survivre.\"],\"DUpGIu\":[\"Fougueux\"],\"RBd2c3\":[\"Pourquoi Delta Green fait-il confiance à cet agent ?\"],\"CyfgYa\":[\"Pourquoi cet agent sert-il Delta Green ?\"],\"imbYbz\":[\"Volonté\"],\"fDORhQ\":[\"Volonté, esprit et stabilité mentale\"],\"kP3ycB\":[\"Volonté\"],\"mYJzKG\":[\"x5\"],\"375wgc\":[\"Vous étendez les connaissances humaines dans un domaine comme la biologie, la physique ou la chimie. Lorsque certaines formes de connaissances provoquent de l'insanité et la mort, il est facile de conclure que certaines hypothèses ne devraient pas être testées.\"],\"vTWySo\":[\"Vous êtes un auteur, un éditeur, un chercheur pour une entreprise ou toutes branches du gouvernement, un bloggeur, un reporter de télévision ou un spécialiste des textes rares. Avec le non naturel, vous avez découvert l'histoire de toute une vie.\"],\"y87741\":[\"Vous reconstituez la vérité à partir de fragments du passé — documents, artefacts et histoires murmurées qui n'ont pas été intégrées aux manuels scolaires. Votre recherche vous a mené sur des chemins que l'académie mainstream refuse d'admettre, dévoilant des motifs dans l'histoire humaine qui suggèrent qu'il existe quelque chose de vaste et de terrible derrière la montée et la chute de la civilisation. Certains archives sont clos pour de bonnes raisons.\"],\"eh3cxY\":[\"Vous dirigez une organisation. Quelqu'un doit sécuriser les fonds, déplacer les ressources et faire des connexions, et c'est vous. Vous contrôlez un budget et êtes responsable de la maintenance de votre programme et du placement de l'argent. Les organisations découvrent les choses les plus surprenantes dans leur quête de profit ou du bien public.\"],\"V1GS2V\":[\"Vous servez et protégez. Les policiers de quartiers patrouillent en uniforme. Les shérifs adjoints répondent à un enforcé élu et ont compétence sur un comté entier. Les détectives arrivent après coup et reconstituent les éléments.\"],\"eYVzTc\":[\"Vous étudiez l'humanité dans toutes ses formes — de tribus isolées à sous-cultures urbaines. Votre travail vous mène dans les coins les plus reculés du monde où des traditions anciennes persistent et où le savoir interdit perdure. Que vous meniez des entrevues en régions en guerre ou que vous analysiez des artefacts culturels qui défient l'explication, vous avez appris que certaines traditions existent pour des raisons plus sombres que personne ne soupçonne.\"],\"doadht\":[\"Vous voyagez dans des terres étranges, rencontrez des personnes intéressantes et essayez de vous entendre avec elles. Les chances sont que vous travaillez pour le Département d'État, bien que USAID, le Service Commercial et le Service des Agriculture Étrangère aient également des FSOs. Quoi qu'il en soit, vous avez eu l'opportunité de connaître des choses exotiques et mortelles ; les choses qui vous qualifient pour la cléarance Delta Green.\"],\"65+vXE\":[\"Vous êtes un membre de l'Armée des États-Unis, de la Marine ou de l'Air Force. Vous avez été formé à la lutte et à la survie. Vous avez vu des choses qui font que le citoyen moyen crie.\"],\"2At5A2\":[\"Your Agent can have up to five Personal Motivations. Add them during character \\n    creation or as your story develops. Each time your Agent experiences a Breaking \\n    Point, remove one motivation to represent their growing trauma.\"],\"blWU4y\":[\"Votre Agent peut avoir jusqu'à cinq motivations personnelles. Ajoutez-les lors de la création du personnage ou au fur et à mesure que votre histoire se développe. Chaque fois que votre Agent subit un point de rupture, retirez une motivation pour représenter leur traumatisme en croissance.\"],\"W0MqKV\":[\"Votre Agent gagne 10% dans la compétence Non naturel et ajoute +20% à l'Occult. Réduisez la SAN de votre Agent de son POW. Votre Agent gagne un nouveau trouble causé par le Non naturel (voir page 72). Réinitialisez le point de rupture de votre Agent à sa nouvelle SAN moins POW.\"],\"1nTbeT\":[\"Votre travail oscille entre la routine de l'entretien de votre matériel, l'excitation lorsque l'alarme finit par venir, et le travail de l'enquête sur un lieu après que le brouillard ait été dissipé. Si vous êtes impliqué avec Delta Green, vous avez clairement trébucheté dans quelque chose de plus grave qu'un incendie de maison.\"],\"+naB54\":[\"Vos outils sont un ordinateur et un smartphone. Vous pourriez déplacer des millions de dollars, ou des bits de données, ou les deux. Ou vous pourriez être un procureur, un avocat de la défense ou un juge.\"],\"cFe/mJ\":[\"Charisme\"]}")as Messages;
