+++ test/test-utils/MockProvider.tsx
@@ -1,7 +1,9 @@
-import { h, ComponentChildren } from 'preact';
+import { ComponentChildren } from 'preact';
 import { ThemeProvider } from 'styled-components';
 import { LocationProvider } from 'preact-iso';
+import { i18n } from '@lingui/core'
 import tokyoDark from 'react95/dist/themes/tokyoDark';
+import { I18nProvider } from '@lingui/react';
 
 interface MockProviderProps {
     children: ComponentChildren;
@@ -12,7 +14,9 @@ function MockProvider ({ children }: MockProviderProps) {
         <div>
             <ThemeProvider theme={tokyoDark}>
                 <LocationProvider>
-                    {children}
+                    <I18nProvider i18n={i18n}>
+                        {children}
+                    </I18nProvider>
                 </LocationProvider>
             </ThemeProvider>
         </div>
